+++ bundles/org.enterprisedomain.classmaker/src/org/enterprisedomain/classmaker/ClassMakerPackage.java
@@ -32,46 +32,43 @@ import org.eclipse.emf.ecore.EReference;
  * <li>and each data type</li>
  * </ul>
  * <!-- end-user-doc -->
- * 
  * @see org.enterprisedomain.classmaker.ClassMakerFactory
  * @model kind="package"
  * @generated
  */
 public interface ClassMakerPackage extends EPackage {
 	/**
-	 * The package name. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The package name.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	String eNAME = "classmaker";
 
 	/**
-	 * The package namespace URI. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The package namespace URI.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	String eNS_URI = "http://org/enterprisedomain/ClassMaker/0.8.28";
 
 	/**
-	 * The package namespace name. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The package namespace name.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	String eNS_PREFIX = "ClassMaker";
 
 	/**
-	 * The singleton instance of the package. <!-- begin-user-doc --> <!--
+	 * The singleton instance of the package.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	ClassMakerPackage eINSTANCE = org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl.init();
 
 	/**
-	 * The meta object id for the
-	 * '{@link org.eclipse.core.runtime.jobs.ISchedulingRule <em>IScheduling
-	 * Rule</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The meta object id for the '{@link org.eclipse.core.runtime.jobs.ISchedulingRule <em>IScheduling Rule</em>}' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @see org.eclipse.core.runtime.jobs.ISchedulingRule
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getISchedulingRule()
 	 * @generated
@@ -81,17 +78,14 @@ public interface ClassMakerPackage extends EPackage {
 	/**
 	 * The number of structural features of the '<em>IScheduling Rule</em>' class.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
 	int ISCHEDULING_RULE_FEATURE_COUNT = 0;
 
 	/**
-	 * The meta object id for the
-	 * '{@link org.enterprisedomain.classmaker.impl.ProjectImpl <em>Project</em>}'
-	 * class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The meta object id for the '{@link org.enterprisedomain.classmaker.impl.ProjectImpl <em>Project</em>}' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @see org.enterprisedomain.classmaker.impl.ProjectImpl
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getProject()
 	 * @generated
@@ -108,9 +102,9 @@ public interface ClassMakerPackage extends EPackage {
 	int PROJECT__MODEL_NAME = ISCHEDULING_RULE_FEATURE_COUNT + 0;
 
 	/**
-	 * The feature id for the '<em><b>Phase</b></em>' attribute. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Phase</b></em>' attribute.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -128,16 +122,15 @@ public interface ClassMakerPackage extends EPackage {
 	/**
 	 * The feature id for the '<em><b>Domain Model</b></em>' containment reference.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
 	int PROJECT__DOMAIN_MODEL = ISCHEDULING_RULE_FEATURE_COUNT + 3;
 
 	/**
-	 * The feature id for the '<em><b>Customizers</b></em>' map. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Customizers</b></em>' map.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -171,9 +164,9 @@ public interface ClassMakerPackage extends EPackage {
 	int PROJECT__PROJECT = ISCHEDULING_RULE_FEATURE_COUNT + 7;
 
 	/**
-	 * The feature id for the '<em><b>Name</b></em>' attribute. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Name</b></em>' attribute.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -198,9 +191,9 @@ public interface ClassMakerPackage extends EPackage {
 	int PROJECT__CHILDREN = ISCHEDULING_RULE_FEATURE_COUNT + 10;
 
 	/**
-	 * The feature id for the '<em><b>Dirty</b></em>' attribute. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Dirty</b></em>' attribute.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -225,18 +218,16 @@ public interface ClassMakerPackage extends EPackage {
 	int PROJECT__RESOURCE_PATH = ISCHEDULING_RULE_FEATURE_COUNT + 13;
 
 	/**
-	 * The feature id for the '<em><b>Need Completion Notification</b></em>'
-	 * attribute. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The feature id for the '<em><b>Need Completion Notification</b></em>' attribute.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 * @ordered
 	 */
 	int PROJECT__NEED_COMPLETION_NOTIFICATION = ISCHEDULING_RULE_FEATURE_COUNT + 14;
 
 	/**
-	 * The feature id for the '<em><b>Completion Notification Adapter</b></em>'
-	 * reference. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The feature id for the '<em><b>Completion Notification Adapter</b></em>' reference.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 * @ordered
 	 */
@@ -245,7 +236,6 @@ public interface ClassMakerPackage extends EPackage {
 	/**
 	 * The feature id for the '<em><b>Resource Reload Listener</b></em>' reference.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -270,9 +260,9 @@ public interface ClassMakerPackage extends EPackage {
 	int PROJECT__REVISION = ISCHEDULING_RULE_FEATURE_COUNT + 18;
 
 	/**
-	 * The feature id for the '<em><b>Revisions</b></em>' map. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Revisions</b></em>' map.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -290,7 +280,6 @@ public interface ClassMakerPackage extends EPackage {
 	/**
 	 * The feature id for the '<em><b>Select Reveal Handler</b></em>' reference.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -306,18 +295,17 @@ public interface ClassMakerPackage extends EPackage {
 	int PROJECT__VERSION = ISCHEDULING_RULE_FEATURE_COUNT + 22;
 
 	/**
-	 * The feature id for the '<em><b>State</b></em>' reference. <!-- begin-user-doc
+	 * The feature id for the '<em><b>State</b></em>' reference.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
 	int PROJECT__STATE = ISCHEDULING_RULE_FEATURE_COUNT + 23;
 
 	/**
-	 * The feature id for the '<em><b>Model Resource Adapter</b></em>' containment
-	 * reference. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The feature id for the '<em><b>Model Resource Adapter</b></em>' containment reference.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 * @ordered
 	 */
@@ -333,10 +321,8 @@ public interface ClassMakerPackage extends EPackage {
 	int PROJECT_FEATURE_COUNT = ISCHEDULING_RULE_FEATURE_COUNT + 25;
 
 	/**
-	 * The meta object id for the
-	 * '{@link org.enterprisedomain.classmaker.impl.ContributionImpl
-	 * <em>Contribution</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The meta object id for the '{@link org.enterprisedomain.classmaker.impl.ContributionImpl <em>Contribution</em>}' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @see org.enterprisedomain.classmaker.impl.ContributionImpl
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getContribution()
 	 * @generated
@@ -353,9 +339,9 @@ public interface ClassMakerPackage extends EPackage {
 	int CONTRIBUTION__MODEL_NAME = PROJECT__MODEL_NAME;
 
 	/**
-	 * The feature id for the '<em><b>Phase</b></em>' attribute. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Phase</b></em>' attribute.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -373,16 +359,15 @@ public interface ClassMakerPackage extends EPackage {
 	/**
 	 * The feature id for the '<em><b>Domain Model</b></em>' containment reference.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
 	int CONTRIBUTION__DOMAIN_MODEL = PROJECT__DOMAIN_MODEL;
 
 	/**
-	 * The feature id for the '<em><b>Customizers</b></em>' map. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Customizers</b></em>' map.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -416,9 +401,9 @@ public interface ClassMakerPackage extends EPackage {
 	int CONTRIBUTION__PROJECT = PROJECT__PROJECT;
 
 	/**
-	 * The feature id for the '<em><b>Name</b></em>' attribute. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Name</b></em>' attribute.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -443,9 +428,9 @@ public interface ClassMakerPackage extends EPackage {
 	int CONTRIBUTION__CHILDREN = PROJECT__CHILDREN;
 
 	/**
-	 * The feature id for the '<em><b>Dirty</b></em>' attribute. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Dirty</b></em>' attribute.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -470,18 +455,16 @@ public interface ClassMakerPackage extends EPackage {
 	int CONTRIBUTION__RESOURCE_PATH = PROJECT__RESOURCE_PATH;
 
 	/**
-	 * The feature id for the '<em><b>Need Completion Notification</b></em>'
-	 * attribute. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The feature id for the '<em><b>Need Completion Notification</b></em>' attribute.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 * @ordered
 	 */
 	int CONTRIBUTION__NEED_COMPLETION_NOTIFICATION = PROJECT__NEED_COMPLETION_NOTIFICATION;
 
 	/**
-	 * The feature id for the '<em><b>Completion Notification Adapter</b></em>'
-	 * reference. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The feature id for the '<em><b>Completion Notification Adapter</b></em>' reference.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 * @ordered
 	 */
@@ -490,7 +473,6 @@ public interface ClassMakerPackage extends EPackage {
 	/**
 	 * The feature id for the '<em><b>Resource Reload Listener</b></em>' reference.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -515,9 +497,9 @@ public interface ClassMakerPackage extends EPackage {
 	int CONTRIBUTION__REVISION = PROJECT__REVISION;
 
 	/**
-	 * The feature id for the '<em><b>Revisions</b></em>' map. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Revisions</b></em>' map.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -535,7 +517,6 @@ public interface ClassMakerPackage extends EPackage {
 	/**
 	 * The feature id for the '<em><b>Select Reveal Handler</b></em>' reference.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -551,18 +532,17 @@ public interface ClassMakerPackage extends EPackage {
 	int CONTRIBUTION__VERSION = PROJECT__VERSION;
 
 	/**
-	 * The feature id for the '<em><b>State</b></em>' reference. <!-- begin-user-doc
+	 * The feature id for the '<em><b>State</b></em>' reference.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
 	int CONTRIBUTION__STATE = PROJECT__STATE;
 
 	/**
-	 * The feature id for the '<em><b>Model Resource Adapter</b></em>' containment
-	 * reference. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The feature id for the '<em><b>Model Resource Adapter</b></em>' containment reference.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 * @ordered
 	 */
@@ -596,10 +576,8 @@ public interface ClassMakerPackage extends EPackage {
 	int CONTRIBUTION_FEATURE_COUNT = PROJECT_FEATURE_COUNT + 2;
 
 	/**
-	 * The meta object id for the
-	 * '{@link org.enterprisedomain.classmaker.impl.ItemImpl <em>Item</em>}' class.
+	 * The meta object id for the '{@link org.enterprisedomain.classmaker.impl.ItemImpl <em>Item</em>}' class.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @see org.enterprisedomain.classmaker.impl.ItemImpl
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getItem()
 	 * @generated
@@ -616,9 +594,9 @@ public interface ClassMakerPackage extends EPackage {
 	int ITEM__MODEL_NAME = 0;
 
 	/**
-	 * The feature id for the '<em><b>Phase</b></em>' attribute. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Phase</b></em>' attribute.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -636,16 +614,15 @@ public interface ClassMakerPackage extends EPackage {
 	/**
 	 * The feature id for the '<em><b>Domain Model</b></em>' containment reference.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
 	int ITEM__DOMAIN_MODEL = 3;
 
 	/**
-	 * The feature id for the '<em><b>Customizers</b></em>' map. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Customizers</b></em>' map.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -688,10 +665,8 @@ public interface ClassMakerPackage extends EPackage {
 	int ITEM_FEATURE_COUNT = 8;
 
 	/**
-	 * The meta object id for the
-	 * '{@link org.enterprisedomain.classmaker.impl.RevisionImpl <em>Revision</em>}'
-	 * class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The meta object id for the '{@link org.enterprisedomain.classmaker.impl.RevisionImpl <em>Revision</em>}' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @see org.enterprisedomain.classmaker.impl.RevisionImpl
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getRevision()
 	 * @generated
@@ -708,9 +683,9 @@ public interface ClassMakerPackage extends EPackage {
 	int REVISION__MODEL_NAME = ITEM__MODEL_NAME;
 
 	/**
-	 * The feature id for the '<em><b>Phase</b></em>' attribute. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Phase</b></em>' attribute.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -728,16 +703,15 @@ public interface ClassMakerPackage extends EPackage {
 	/**
 	 * The feature id for the '<em><b>Domain Model</b></em>' containment reference.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
 	int REVISION__DOMAIN_MODEL = ITEM__DOMAIN_MODEL;
 
 	/**
-	 * The feature id for the '<em><b>Customizers</b></em>' map. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Customizers</b></em>' map.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -771,9 +745,9 @@ public interface ClassMakerPackage extends EPackage {
 	int REVISION__PROJECT = ITEM__PROJECT;
 
 	/**
-	 * The feature id for the '<em><b>State</b></em>' reference. <!-- begin-user-doc
+	 * The feature id for the '<em><b>State</b></em>' reference.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -825,10 +799,8 @@ public interface ClassMakerPackage extends EPackage {
 	int REVISION_FEATURE_COUNT = ITEM_FEATURE_COUNT + 5;
 
 	/**
-	 * The meta object id for the
-	 * '{@link org.enterprisedomain.classmaker.impl.StateImpl <em>State</em>}'
-	 * class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The meta object id for the '{@link org.enterprisedomain.classmaker.impl.StateImpl <em>State</em>}' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @see org.enterprisedomain.classmaker.impl.StateImpl
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getState()
 	 * @generated
@@ -845,9 +817,9 @@ public interface ClassMakerPackage extends EPackage {
 	int STATE__MODEL_NAME = ITEM__MODEL_NAME;
 
 	/**
-	 * The feature id for the '<em><b>Phase</b></em>' attribute. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Phase</b></em>' attribute.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -865,16 +837,15 @@ public interface ClassMakerPackage extends EPackage {
 	/**
 	 * The feature id for the '<em><b>Domain Model</b></em>' containment reference.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
 	int STATE__DOMAIN_MODEL = ITEM__DOMAIN_MODEL;
 
 	/**
-	 * The feature id for the '<em><b>Customizers</b></em>' map. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Customizers</b></em>' map.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -919,7 +890,6 @@ public interface ClassMakerPackage extends EPackage {
 	/**
 	 * The feature id for the '<em><b>Edit Plugin Class Name</b></em>' attribute.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -928,7 +898,6 @@ public interface ClassMakerPackage extends EPackage {
 	/**
 	 * The feature id for the '<em><b>Editor Plugin Class Name</b></em>' attribute.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -937,7 +906,6 @@ public interface ClassMakerPackage extends EPackage {
 	/**
 	 * The feature id for the '<em><b>Required Plugins</b></em>' attribute list.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -973,16 +941,14 @@ public interface ClassMakerPackage extends EPackage {
 	/**
 	 * The feature id for the '<em><b>Edit Deployable Unit Name</b></em>' attribute.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
 	int STATE__EDIT_DEPLOYABLE_UNIT_NAME = ITEM_FEATURE_COUNT + 7;
 
 	/**
-	 * The feature id for the '<em><b>Editor Deployable Unit Name</b></em>'
-	 * attribute. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The feature id for the '<em><b>Editor Deployable Unit Name</b></em>' attribute.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 * @ordered
 	 */
@@ -1052,9 +1018,9 @@ public interface ClassMakerPackage extends EPackage {
 	int STATE__MAKING = ITEM_FEATURE_COUNT + 15;
 
 	/**
-	 * The feature id for the '<em><b>Edit</b></em>' attribute. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Edit</b></em>' attribute.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -1088,10 +1054,8 @@ public interface ClassMakerPackage extends EPackage {
 	int STATE_FEATURE_COUNT = ITEM_FEATURE_COUNT + 19;
 
 	/**
-	 * The meta object id for the
-	 * '{@link org.enterprisedomain.classmaker.impl.StrategyImpl <em>Strategy</em>}'
-	 * class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The meta object id for the '{@link org.enterprisedomain.classmaker.impl.StrategyImpl <em>Strategy</em>}' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @see org.enterprisedomain.classmaker.impl.StrategyImpl
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getStrategy()
 	 * @generated
@@ -1135,9 +1099,9 @@ public interface ClassMakerPackage extends EPackage {
 	int STRATEGY__LOADERS = 3;
 
 	/**
-	 * The feature id for the '<em><b>State</b></em>' reference. <!-- begin-user-doc
+	 * The feature id for the '<em><b>State</b></em>' reference.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -1153,9 +1117,8 @@ public interface ClassMakerPackage extends EPackage {
 	int STRATEGY_FEATURE_COUNT = 5;
 
 	/**
-	 * The meta object id for the '{@link org.eclipse.emf.common.notify.Adapter
-	 * <em>Adapter</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The meta object id for the '{@link org.eclipse.emf.common.notify.Adapter <em>Adapter</em>}' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @see org.eclipse.emf.common.notify.Adapter
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getAdapter()
 	 * @generated
@@ -1172,10 +1135,8 @@ public interface ClassMakerPackage extends EPackage {
 	int ADAPTER_FEATURE_COUNT = 0;
 
 	/**
-	 * The meta object id for the
-	 * '{@link org.enterprisedomain.classmaker.impl.WorkspaceImpl
-	 * <em>Workspace</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The meta object id for the '{@link org.enterprisedomain.classmaker.impl.WorkspaceImpl <em>Workspace</em>}' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @see org.enterprisedomain.classmaker.impl.WorkspaceImpl
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getWorkspace()
 	 * @generated
@@ -1185,7 +1146,6 @@ public interface ClassMakerPackage extends EPackage {
 	/**
 	 * The feature id for the '<em><b>Projects</b></em>' containment reference list.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -1201,9 +1161,9 @@ public interface ClassMakerPackage extends EPackage {
 	int WORKSPACE__RESOURCE_SET = ISCHEDULING_RULE_FEATURE_COUNT + 1;
 
 	/**
-	 * The feature id for the '<em><b>Customizers</b></em>' map. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Customizers</b></em>' map.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -1249,9 +1209,9 @@ public interface ClassMakerPackage extends EPackage {
 	int LONG_TO_STATE_MAP_ENTRY = 7;
 
 	/**
-	 * The feature id for the '<em><b>Key</b></em>' attribute. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Key</b></em>' attribute.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -1267,20 +1227,17 @@ public interface ClassMakerPackage extends EPackage {
 	int LONG_TO_STATE_MAP_ENTRY__VALUE = 1;
 
 	/**
-	 * The number of structural features of the '<em>Long To State Map Entry</em>'
-	 * class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The number of structural features of the '<em>Long To State Map Entry</em>' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 * @ordered
 	 */
 	int LONG_TO_STATE_MAP_ENTRY_FEATURE_COUNT = 2;
 
 	/**
-	 * The meta object id for the
-	 * '{@link org.enterprisedomain.classmaker.impl.VersionToRevisionMapEntryImpl
-	 * <em>Version To Revision Map Entry</em>}' class. <!-- begin-user-doc --> <!--
+	 * The meta object id for the '{@link org.enterprisedomain.classmaker.impl.VersionToRevisionMapEntryImpl <em>Version To Revision Map Entry</em>}' class.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @see org.enterprisedomain.classmaker.impl.VersionToRevisionMapEntryImpl
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getVersionToRevisionMapEntry()
 	 * @generated
@@ -1288,9 +1245,9 @@ public interface ClassMakerPackage extends EPackage {
 	int VERSION_TO_REVISION_MAP_ENTRY = 8;
 
 	/**
-	 * The feature id for the '<em><b>Key</b></em>' attribute. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Key</b></em>' attribute.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -1306,18 +1263,16 @@ public interface ClassMakerPackage extends EPackage {
 	int VERSION_TO_REVISION_MAP_ENTRY__VALUE = 1;
 
 	/**
-	 * The number of structural features of the '<em>Version To Revision Map
-	 * Entry</em>' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The number of structural features of the '<em>Version To Revision Map Entry</em>' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 * @ordered
 	 */
 	int VERSION_TO_REVISION_MAP_ENTRY_FEATURE_COUNT = 2;
 
 	/**
-	 * The meta object id for the '{@link java.util.concurrent.Executor
-	 * <em>Executor</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The meta object id for the '{@link java.util.concurrent.Executor <em>Executor</em>}' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @see java.util.concurrent.Executor
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getExecutor()
 	 * @generated
@@ -1334,9 +1289,8 @@ public interface ClassMakerPackage extends EPackage {
 	int EXECUTOR_FEATURE_COUNT = 0;
 
 	/**
-	 * The meta object id for the '{@link java.util.concurrent.Future
-	 * <em>Future</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The meta object id for the '{@link java.util.concurrent.Future <em>Future</em>}' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @see java.util.concurrent.Future
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getFuture()
 	 * @generated
@@ -1366,17 +1320,14 @@ public interface ClassMakerPackage extends EPackage {
 	/**
 	 * The number of structural features of the '<em>IAdapter Factory</em>' class.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
 	int IADAPTER_FACTORY_FEATURE_COUNT = 0;
 
 	/**
-	 * The meta object id for the
-	 * '{@link org.enterprisedomain.classmaker.impl.CustomizerImpl
-	 * <em>Customizer</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The meta object id for the '{@link org.enterprisedomain.classmaker.impl.CustomizerImpl <em>Customizer</em>}' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @see org.enterprisedomain.classmaker.impl.CustomizerImpl
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getCustomizer()
 	 * @generated
@@ -1384,9 +1335,9 @@ public interface ClassMakerPackage extends EPackage {
 	int CUSTOMIZER = 13;
 
 	/**
-	 * The feature id for the '<em><b>Rank</b></em>' attribute. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Rank</b></em>' attribute.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -1402,10 +1353,8 @@ public interface ClassMakerPackage extends EPackage {
 	int CUSTOMIZER_FEATURE_COUNT = 1;
 
 	/**
-	 * The meta object id for the
-	 * '{@link org.enterprisedomain.classmaker.impl.StageQualifierImpl <em>Stage
-	 * Qualifier</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The meta object id for the '{@link org.enterprisedomain.classmaker.impl.StageQualifierImpl <em>Stage Qualifier</em>}' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @see org.enterprisedomain.classmaker.impl.StageQualifierImpl
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getStageQualifier()
 	 * @generated
@@ -1413,18 +1362,18 @@ public interface ClassMakerPackage extends EPackage {
 	int STAGE_QUALIFIER = 14;
 
 	/**
-	 * The feature id for the '<em><b>Stage</b></em>' attribute. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Stage</b></em>' attribute.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
 	int STAGE_QUALIFIER__STAGE = 0;
 
 	/**
-	 * The feature id for the '<em><b>Step</b></em>' attribute. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Step</b></em>' attribute.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -1433,18 +1382,15 @@ public interface ClassMakerPackage extends EPackage {
 	/**
 	 * The number of structural features of the '<em>Stage Qualifier</em>' class.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
 	int STAGE_QUALIFIER_FEATURE_COUNT = 2;
 
 	/**
-	 * The meta object id for the
-	 * '{@link org.enterprisedomain.classmaker.impl.StageQualifierToCustomizerMapEntryImpl
-	 * <em>Stage Qualifier To Customizer Map Entry</em>}' class. <!-- begin-user-doc
+	 * The meta object id for the '{@link org.enterprisedomain.classmaker.impl.StageQualifierToCustomizerMapEntryImpl <em>Stage Qualifier To Customizer Map Entry</em>}' class.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @see org.enterprisedomain.classmaker.impl.StageQualifierToCustomizerMapEntryImpl
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getStageQualifierToCustomizerMapEntry()
 	 * @generated
@@ -1452,9 +1398,9 @@ public interface ClassMakerPackage extends EPackage {
 	int STAGE_QUALIFIER_TO_CUSTOMIZER_MAP_ENTRY = 15;
 
 	/**
-	 * The feature id for the '<em><b>Key</b></em>' reference. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Key</b></em>' reference.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -1470,19 +1416,16 @@ public interface ClassMakerPackage extends EPackage {
 	int STAGE_QUALIFIER_TO_CUSTOMIZER_MAP_ENTRY__VALUE = 1;
 
 	/**
-	 * The number of structural features of the '<em>Stage Qualifier To Customizer
-	 * Map Entry</em>' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The number of structural features of the '<em>Stage Qualifier To Customizer Map Entry</em>' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 * @ordered
 	 */
 	int STAGE_QUALIFIER_TO_CUSTOMIZER_MAP_ENTRY_FEATURE_COUNT = 2;
 
 	/**
-	 * The meta object id for the
-	 * '{@link org.enterprisedomain.classmaker.impl.ModelPairImpl <em>Model
-	 * Pair</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The meta object id for the '{@link org.enterprisedomain.classmaker.impl.ModelPairImpl <em>Model Pair</em>}' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @see org.enterprisedomain.classmaker.impl.ModelPairImpl
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getModelPair()
 	 * @generated
@@ -1526,10 +1469,8 @@ public interface ClassMakerPackage extends EPackage {
 	int MODEL_PAIR_FEATURE_COUNT = 3;
 
 	/**
-	 * The meta object id for the
-	 * '{@link org.enterprisedomain.classmaker.impl.SCMOperatorImpl <em>SCM
-	 * Operator</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The meta object id for the '{@link org.enterprisedomain.classmaker.impl.SCMOperatorImpl <em>SCM Operator</em>}' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @see org.enterprisedomain.classmaker.impl.SCMOperatorImpl
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getSCMOperator()
 	 * @generated
@@ -1564,10 +1505,8 @@ public interface ClassMakerPackage extends EPackage {
 	int SCM_OPERATOR_FEATURE_COUNT = 2;
 
 	/**
-	 * The meta object id for the
-	 * '{@link org.enterprisedomain.classmaker.impl.SCMRegistryImpl <em>SCM
-	 * Registry</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The meta object id for the '{@link org.enterprisedomain.classmaker.impl.SCMRegistryImpl <em>SCM Registry</em>}' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @see org.enterprisedomain.classmaker.impl.SCMRegistryImpl
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getSCMRegistry()
 	 * @generated
@@ -1584,9 +1523,8 @@ public interface ClassMakerPackage extends EPackage {
 	int SCM_REGISTRY_FEATURE_COUNT = 0;
 
 	/**
-	 * The meta object id for the '{@link org.eclipse.emf.ecore.resource.Resource
-	 * <em>Resource</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The meta object id for the '{@link org.eclipse.emf.ecore.resource.Resource <em>Resource</em>}' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @see org.eclipse.emf.ecore.resource.Resource
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getResource()
 	 * @generated
@@ -1603,11 +1541,9 @@ public interface ClassMakerPackage extends EPackage {
 	int RESOURCE_FEATURE_COUNT = 0;
 
 	/**
-	 * The meta object id for the
-	 * '{@link org.enterprisedomain.classmaker.impl.ResourceChangeListenerImpl
-	 * <em>Resource Change Listener</em>}' class. <!-- begin-user-doc --> <!--
+	 * The meta object id for the '{@link org.enterprisedomain.classmaker.impl.ResourceChangeListenerImpl <em>Resource Change Listener</em>}' class.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @see org.enterprisedomain.classmaker.impl.ResourceChangeListenerImpl
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getResourceChangeListener()
 	 * @generated
@@ -1615,18 +1551,16 @@ public interface ClassMakerPackage extends EPackage {
 	int RESOURCE_CHANGE_LISTENER = 21;
 
 	/**
-	 * The number of structural features of the '<em>Resource Change Listener</em>'
-	 * class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The number of structural features of the '<em>Resource Change Listener</em>' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 * @ordered
 	 */
 	int RESOURCE_CHANGE_LISTENER_FEATURE_COUNT = 0;
 
 	/**
-	 * The meta object id for the '{@link org.eclipse.emf.common.notify.Notification
-	 * <em>Notification</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The meta object id for the '{@link org.eclipse.emf.common.notify.Notification <em>Notification</em>}' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @see org.eclipse.emf.common.notify.Notification
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getNotification()
 	 * @generated
@@ -1643,11 +1577,9 @@ public interface ClassMakerPackage extends EPackage {
 	int NOTIFICATION_FEATURE_COUNT = 0;
 
 	/**
-	 * The meta object id for the
-	 * '{@link org.enterprisedomain.classmaker.impl.CompletionListenerImpl
-	 * <em>Completion Listener</em>}' class. <!-- begin-user-doc --> <!--
+	 * The meta object id for the '{@link org.enterprisedomain.classmaker.impl.CompletionListenerImpl <em>Completion Listener</em>}' class.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @see org.enterprisedomain.classmaker.impl.CompletionListenerImpl
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getCompletionListener()
 	 * @generated
@@ -1655,19 +1587,16 @@ public interface ClassMakerPackage extends EPackage {
 	int COMPLETION_LISTENER = 23;
 
 	/**
-	 * The number of structural features of the '<em>Completion Listener</em>'
-	 * class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The number of structural features of the '<em>Completion Listener</em>' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 * @ordered
 	 */
 	int COMPLETION_LISTENER_FEATURE_COUNT = 0;
 
 	/**
-	 * The meta object id for the
-	 * '{@link org.enterprisedomain.classmaker.jobs.Worker <em>Worker</em>}' class.
+	 * The meta object id for the '{@link org.enterprisedomain.classmaker.jobs.Worker <em>Worker</em>}' class.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @see org.enterprisedomain.classmaker.jobs.Worker
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getWorker()
 	 * @generated
@@ -1675,10 +1604,8 @@ public interface ClassMakerPackage extends EPackage {
 	int WORKER = 28;
 
 	/**
-	 * The meta object id for the
-	 * '{@link org.enterprisedomain.classmaker.impl.ResourceAdapterImpl <em>Resource
-	 * Adapter</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The meta object id for the '{@link org.enterprisedomain.classmaker.impl.ResourceAdapterImpl <em>Resource Adapter</em>}' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @see org.enterprisedomain.classmaker.impl.ResourceAdapterImpl
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getResourceAdapter()
 	 * @generated
@@ -1715,17 +1642,14 @@ public interface ClassMakerPackage extends EPackage {
 	/**
 	 * The number of structural features of the '<em>Resource Adapter</em>' class.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
 	int RESOURCE_ADAPTER_FEATURE_COUNT = ADAPTER_FEATURE_COUNT + 3;
 
 	/**
-	 * The meta object id for the
-	 * '{@link org.enterprisedomain.classmaker.impl.ClassMakerServiceImpl
-	 * <em>Service</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The meta object id for the '{@link org.enterprisedomain.classmaker.impl.ClassMakerServiceImpl <em>Service</em>}' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerServiceImpl
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getClassMakerService()
 	 * @generated
@@ -1751,10 +1675,8 @@ public interface ClassMakerPackage extends EPackage {
 	int CLASS_MAKER_SERVICE_FEATURE_COUNT = 1;
 
 	/**
-	 * The meta object id for the
-	 * '{@link org.enterprisedomain.classmaker.impl.BlueprintImpl
-	 * <em>Blueprint</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The meta object id for the '{@link org.enterprisedomain.classmaker.impl.BlueprintImpl <em>Blueprint</em>}' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @see org.enterprisedomain.classmaker.impl.BlueprintImpl
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getBlueprint()
 	 * @generated
@@ -1782,16 +1704,15 @@ public interface ClassMakerPackage extends EPackage {
 	/**
 	 * The feature id for the '<em><b>Completion Listeners</b></em>' reference list.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
 	int BLUEPRINT__COMPLETION_LISTENERS = 2;
 
 	/**
-	 * The feature id for the '<em><b>Edit</b></em>' attribute. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Edit</b></em>' attribute.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -1816,11 +1737,9 @@ public interface ClassMakerPackage extends EPackage {
 	int BLUEPRINT_FEATURE_COUNT = 5;
 
 	/**
-	 * The meta object id for the
-	 * '{@link org.enterprisedomain.classmaker.impl.CompletionNotificationAdapterImpl
-	 * <em>Completion Notification Adapter</em>}' class. <!-- begin-user-doc -->
+	 * The meta object id for the '{@link org.enterprisedomain.classmaker.impl.CompletionNotificationAdapterImpl <em>Completion Notification Adapter</em>}' class.
+	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
-	 * 
 	 * @see org.enterprisedomain.classmaker.impl.CompletionNotificationAdapterImpl
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getCompletionNotificationAdapter()
 	 * @generated
@@ -1828,9 +1747,9 @@ public interface ClassMakerPackage extends EPackage {
 	int COMPLETION_NOTIFICATION_ADAPTER = 27;
 
 	/**
-	 * The feature id for the '<em><b>Error</b></em>' attribute. <!-- begin-user-doc
+	 * The feature id for the '<em><b>Error</b></em>' attribute.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 * @ordered
 	 */
@@ -1846,9 +1765,8 @@ public interface ClassMakerPackage extends EPackage {
 	int COMPLETION_NOTIFICATION_ADAPTER__PROJECT = ADAPTER_FEATURE_COUNT + 1;
 
 	/**
-	 * The number of structural features of the '<em>Completion Notification
-	 * Adapter</em>' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The number of structural features of the '<em>Completion Notification Adapter</em>' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 * @ordered
 	 */
@@ -1864,11 +1782,9 @@ public interface ClassMakerPackage extends EPackage {
 	int WORKER_FEATURE_COUNT = 0;
 
 	/**
-	 * The meta object id for the
-	 * '{@link org.enterprisedomain.classmaker.impl.SelectRevealHandlerImpl
-	 * <em>Select Reveal Handler</em>}' class. <!-- begin-user-doc --> <!--
+	 * The meta object id for the '{@link org.enterprisedomain.classmaker.impl.SelectRevealHandlerImpl <em>Select Reveal Handler</em>}' class.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @see org.enterprisedomain.classmaker.impl.SelectRevealHandlerImpl
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getSelectRevealHandler()
 	 * @generated
@@ -1876,18 +1792,17 @@ public interface ClassMakerPackage extends EPackage {
 	int SELECT_REVEAL_HANDLER = 29;
 
 	/**
-	 * The number of structural features of the '<em>Select Reveal Handler</em>'
-	 * class. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The number of structural features of the '<em>Select Reveal Handler</em>' class.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 * @ordered
 	 */
 	int SELECT_REVEAL_HANDLER_FEATURE_COUNT = 0;
 
 	/**
-	 * The meta object id for the '<em>Locale</em>' data type. <!-- begin-user-doc
+	 * The meta object id for the '<em>Locale</em>' data type.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @see java.util.Locale
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getLocale()
 	 * @generated
@@ -1895,9 +1810,8 @@ public interface ClassMakerPackage extends EPackage {
 	int LOCALE = 41;
 
 	/**
-	 * The meta object id for the '{@link org.enterprisedomain.classmaker.Stage
-	 * <em>Stage</em>}' enum. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * The meta object id for the '{@link org.enterprisedomain.classmaker.Stage <em>Stage</em>}' enum.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @see org.enterprisedomain.classmaker.Stage
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getStage()
 	 * @generated
@@ -1957,7 +1871,6 @@ public interface ClassMakerPackage extends EPackage {
 	/**
 	 * The meta object id for the '<em>Invocation Target Exception</em>' data type.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @see java.lang.reflect.InvocationTargetException
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getInvocationTargetException()
 	 * @generated
@@ -1965,9 +1878,9 @@ public interface ClassMakerPackage extends EPackage {
 	int INVOCATION_TARGET_EXCEPTION = 36;
 
 	/**
-	 * The meta object id for the '<em>IStatus</em>' data type. <!-- begin-user-doc
+	 * The meta object id for the '<em>IStatus</em>' data type.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @see org.eclipse.core.runtime.IStatus
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getIStatus()
 	 * @generated
@@ -1975,9 +1888,9 @@ public interface ClassMakerPackage extends EPackage {
 	int ISTATUS = 37;
 
 	/**
-	 * The meta object id for the '<em>URI</em>' data type. <!-- begin-user-doc -->
+	 * The meta object id for the '<em>URI</em>' data type.
+	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
-	 * 
 	 * @see org.eclipse.emf.common.util.URI
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getURI()
 	 * @generated
@@ -1985,9 +1898,9 @@ public interface ClassMakerPackage extends EPackage {
 	int URI = 38;
 
 	/**
-	 * The meta object id for the '<em>Name</em>' data type. <!-- begin-user-doc -->
+	 * The meta object id for the '<em>Name</em>' data type.
+	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
-	 * 
 	 * @see java.lang.String
 	 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getName_()
 	 * @generated
@@ -2005,10 +1918,8 @@ public interface ClassMakerPackage extends EPackage {
 	int EXCEPTION = 40;
 
 	/**
-	 * Returns the meta object for class
-	 * '{@link org.enterprisedomain.classmaker.Contribution <em>Contribution</em>}'.
+	 * Returns the meta object for class '{@link org.enterprisedomain.classmaker.Contribution <em>Contribution</em>}'.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @return the meta object for class '<em>Contribution</em>'.
 	 * @see org.enterprisedomain.classmaker.Contribution
 	 * @generated
@@ -2016,10 +1927,8 @@ public interface ClassMakerPackage extends EPackage {
 	EClass getContribution();
 
 	/**
-	 * Returns the meta object for the attribute list
-	 * '{@link org.enterprisedomain.classmaker.Contribution#getDependencies
-	 * <em>Dependencies</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute list '{@link org.enterprisedomain.classmaker.Contribution#getDependencies <em>Dependencies</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute list '<em>Dependencies</em>'.
 	 * @see org.enterprisedomain.classmaker.Contribution#getDependencies()
 	 * @see #getContribution()
@@ -2028,10 +1937,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getContribution_Dependencies();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.Contribution#getLatestVersion
-	 * <em>Latest Version</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.Contribution#getLatestVersion <em>Latest Version</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute '<em>Latest Version</em>'.
 	 * @see org.enterprisedomain.classmaker.Contribution#getLatestVersion()
 	 * @see #getContribution()
@@ -2051,10 +1958,8 @@ public interface ClassMakerPackage extends EPackage {
 	EClass getRevision();
 
 	/**
-	 * Returns the meta object for the reference
-	 * '{@link org.enterprisedomain.classmaker.Revision#getState <em>State</em>}'.
+	 * Returns the meta object for the reference '{@link org.enterprisedomain.classmaker.Revision#getState <em>State</em>}'.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @return the meta object for the reference '<em>State</em>'.
 	 * @see org.enterprisedomain.classmaker.Revision#getState()
 	 * @see #getRevision()
@@ -2063,10 +1968,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getRevision_State();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.Revision#getTimestamp
-	 * <em>Timestamp</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.Revision#getTimestamp <em>Timestamp</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute '<em>Timestamp</em>'.
 	 * @see org.enterprisedomain.classmaker.Revision#getTimestamp()
 	 * @see #getRevision()
@@ -2075,10 +1978,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getRevision_Timestamp();
 
 	/**
-	 * Returns the meta object for the map
-	 * '{@link org.enterprisedomain.classmaker.Revision#getStateHistory <em>State
-	 * History</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the map '{@link org.enterprisedomain.classmaker.Revision#getStateHistory <em>State History</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the map '<em>State History</em>'.
 	 * @see org.enterprisedomain.classmaker.Revision#getStateHistory()
 	 * @see #getRevision()
@@ -2087,10 +1988,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getRevision_StateHistory();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.Revision#getLatestTimestamp
-	 * <em>Latest Timestamp</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.Revision#getLatestTimestamp <em>Latest Timestamp</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute '<em>Latest Timestamp</em>'.
 	 * @see org.enterprisedomain.classmaker.Revision#getLatestTimestamp()
 	 * @see #getRevision()
@@ -2099,10 +1998,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getRevision_LatestTimestamp();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.Revision#getVersion
-	 * <em>Version</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.Revision#getVersion <em>Version</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute '<em>Version</em>'.
 	 * @see org.enterprisedomain.classmaker.Revision#getVersion()
 	 * @see #getRevision()
@@ -2122,10 +2019,8 @@ public interface ClassMakerPackage extends EPackage {
 	EClass getState();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.State#getPackageClassName <em>Package
-	 * Class Name</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.State#getPackageClassName <em>Package Class Name</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute '<em>Package Class Name</em>'.
 	 * @see org.enterprisedomain.classmaker.State#getPackageClassName()
 	 * @see #getState()
@@ -2134,10 +2029,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getState_PackageClassName();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.State#getEditPluginClassName <em>Edit
-	 * Plugin Class Name</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.State#getEditPluginClassName <em>Edit Plugin Class Name</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute '<em>Edit Plugin Class Name</em>'.
 	 * @see org.enterprisedomain.classmaker.State#getEditPluginClassName()
 	 * @see #getState()
@@ -2146,13 +2039,10 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getState_EditPluginClassName();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.State#getEditorPluginClassName
-	 * <em>Editor Plugin Class Name</em>}'. <!-- begin-user-doc --> <!--
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.State#getEditorPluginClassName <em>Editor Plugin Class Name</em>}'.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
-	 * @return the meta object for the attribute '<em>Editor Plugin Class
-	 *         Name</em>'.
+	 * @return the meta object for the attribute '<em>Editor Plugin Class Name</em>'.
 	 * @see org.enterprisedomain.classmaker.State#getEditorPluginClassName()
 	 * @see #getState()
 	 * @generated
@@ -2160,10 +2050,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getState_EditorPluginClassName();
 
 	/**
-	 * Returns the meta object for the attribute list
-	 * '{@link org.enterprisedomain.classmaker.State#getRequiredPlugins <em>Required
-	 * Plugins</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute list '{@link org.enterprisedomain.classmaker.State#getRequiredPlugins <em>Required Plugins</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute list '<em>Required Plugins</em>'.
 	 * @see org.enterprisedomain.classmaker.State#getRequiredPlugins()
 	 * @see #getState()
@@ -2172,10 +2060,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getState_RequiredPlugins();
 
 	/**
-	 * Returns the meta object for the reference
-	 * '{@link org.enterprisedomain.classmaker.State#getRevision
-	 * <em>Revision</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the reference '{@link org.enterprisedomain.classmaker.State#getRevision <em>Revision</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the reference '<em>Revision</em>'.
 	 * @see org.enterprisedomain.classmaker.State#getRevision()
 	 * @see #getState()
@@ -2184,10 +2070,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getState_Revision();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.State#getTimestamp
-	 * <em>Timestamp</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.State#getTimestamp <em>Timestamp</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute '<em>Timestamp</em>'.
 	 * @see org.enterprisedomain.classmaker.State#getTimestamp()
 	 * @see #getState()
@@ -2209,13 +2093,10 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getState_DeployableUnitName();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.State#getEditDeployableUnitName
-	 * <em>Edit Deployable Unit Name</em>}'. <!-- begin-user-doc --> <!--
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.State#getEditDeployableUnitName <em>Edit Deployable Unit Name</em>}'.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
-	 * @return the meta object for the attribute '<em>Edit Deployable Unit
-	 *         Name</em>'.
+	 * @return the meta object for the attribute '<em>Edit Deployable Unit Name</em>'.
 	 * @see org.enterprisedomain.classmaker.State#getEditDeployableUnitName()
 	 * @see #getState()
 	 * @generated
@@ -2223,13 +2104,10 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getState_EditDeployableUnitName();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.State#getEditorDeployableUnitName
-	 * <em>Editor Deployable Unit Name</em>}'. <!-- begin-user-doc --> <!--
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.State#getEditorDeployableUnitName <em>Editor Deployable Unit Name</em>}'.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
-	 * @return the meta object for the attribute '<em>Editor Deployable Unit
-	 *         Name</em>'.
+	 * @return the meta object for the attribute '<em>Editor Deployable Unit Name</em>'.
 	 * @see org.enterprisedomain.classmaker.State#getEditorDeployableUnitName()
 	 * @see #getState()
 	 * @generated
@@ -2237,10 +2115,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getState_EditorDeployableUnitName();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.State#getJobFamily <em>Job
-	 * Family</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.State#getJobFamily <em>Job Family</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute '<em>Job Family</em>'.
 	 * @see org.enterprisedomain.classmaker.State#getJobFamily()
 	 * @see #getState()
@@ -2249,10 +2125,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getState_JobFamily();
 
 	/**
-	 * Returns the meta object for the reference
-	 * '{@link org.enterprisedomain.classmaker.State#getResource
-	 * <em>Resource</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the reference '{@link org.enterprisedomain.classmaker.State#getResource <em>Resource</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the reference '<em>Resource</em>'.
 	 * @see org.enterprisedomain.classmaker.State#getResource()
 	 * @see #getState()
@@ -2261,10 +2135,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getState_Resource();
 
 	/**
-	 * Returns the meta object for the attribute list
-	 * '{@link org.enterprisedomain.classmaker.State#getCommitIds <em>Commit
-	 * Ids</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute list '{@link org.enterprisedomain.classmaker.State#getCommitIds <em>Commit Ids</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute list '<em>Commit Ids</em>'.
 	 * @see org.enterprisedomain.classmaker.State#getCommitIds()
 	 * @see #getState()
@@ -2273,10 +2145,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getState_CommitIds();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.State#getCommitId <em>Commit
-	 * Id</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.State#getCommitId <em>Commit Id</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute '<em>Commit Id</em>'.
 	 * @see org.enterprisedomain.classmaker.State#getCommitId()
 	 * @see #getState()
@@ -2285,10 +2155,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getState_CommitId();
 
 	/**
-	 * Returns the meta object for the map
-	 * '{@link org.enterprisedomain.classmaker.State#getStateCustomizers <em>State
-	 * Customizers</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the map '{@link org.enterprisedomain.classmaker.State#getStateCustomizers <em>State Customizers</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the map '<em>State Customizers</em>'.
 	 * @see org.enterprisedomain.classmaker.State#getStateCustomizers()
 	 * @see #getState()
@@ -2297,10 +2165,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getState_StateCustomizers();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.State#getProjectName <em>Project
-	 * Name</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.State#getProjectName <em>Project Name</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute '<em>Project Name</em>'.
 	 * @see org.enterprisedomain.classmaker.State#getProjectName()
 	 * @see #getState()
@@ -2309,10 +2175,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getState_ProjectName();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.State#isMaking <em>Making</em>}'.
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.State#isMaking <em>Making</em>}'.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @return the meta object for the attribute '<em>Making</em>'.
 	 * @see org.enterprisedomain.classmaker.State#isMaking()
 	 * @see #getState()
@@ -2333,10 +2197,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getState_Edit();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.State#isEditor <em>Editor</em>}'.
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.State#isEditor <em>Editor</em>}'.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @return the meta object for the attribute '<em>Editor</em>'.
 	 * @see org.enterprisedomain.classmaker.State#isEditor()
 	 * @see #getState()
@@ -2345,10 +2207,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getState_Editor();
 
 	/**
-	 * Returns the meta object for the reference
-	 * '{@link org.enterprisedomain.classmaker.State#getStrategy
-	 * <em>Strategy</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the reference '{@link org.enterprisedomain.classmaker.State#getStrategy <em>Strategy</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the reference '<em>Strategy</em>'.
 	 * @see org.enterprisedomain.classmaker.State#getStrategy()
 	 * @see #getState()
@@ -2368,10 +2228,8 @@ public interface ClassMakerPackage extends EPackage {
 	EClass getStrategy();
 
 	/**
-	 * Returns the meta object for the reference list
-	 * '{@link org.enterprisedomain.classmaker.Strategy#getGenerators
-	 * <em>Generators</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the reference list '{@link org.enterprisedomain.classmaker.Strategy#getGenerators <em>Generators</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the reference list '<em>Generators</em>'.
 	 * @see org.enterprisedomain.classmaker.Strategy#getGenerators()
 	 * @see #getStrategy()
@@ -2380,10 +2238,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getStrategy_Generators();
 
 	/**
-	 * Returns the meta object for the reference list
-	 * '{@link org.enterprisedomain.classmaker.Strategy#getExporters
-	 * <em>Exporters</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the reference list '{@link org.enterprisedomain.classmaker.Strategy#getExporters <em>Exporters</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the reference list '<em>Exporters</em>'.
 	 * @see org.enterprisedomain.classmaker.Strategy#getExporters()
 	 * @see #getStrategy()
@@ -2392,10 +2248,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getStrategy_Exporters();
 
 	/**
-	 * Returns the meta object for the reference list
-	 * '{@link org.enterprisedomain.classmaker.Strategy#getInstallers
-	 * <em>Installers</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the reference list '{@link org.enterprisedomain.classmaker.Strategy#getInstallers <em>Installers</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the reference list '<em>Installers</em>'.
 	 * @see org.enterprisedomain.classmaker.Strategy#getInstallers()
 	 * @see #getStrategy()
@@ -2404,10 +2258,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getStrategy_Installers();
 
 	/**
-	 * Returns the meta object for the reference list
-	 * '{@link org.enterprisedomain.classmaker.Strategy#getLoaders
-	 * <em>Loaders</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the reference list '{@link org.enterprisedomain.classmaker.Strategy#getLoaders <em>Loaders</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the reference list '<em>Loaders</em>'.
 	 * @see org.enterprisedomain.classmaker.Strategy#getLoaders()
 	 * @see #getStrategy()
@@ -2416,10 +2268,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getStrategy_Loaders();
 
 	/**
-	 * Returns the meta object for the reference
-	 * '{@link org.enterprisedomain.classmaker.Strategy#getState <em>State</em>}'.
+	 * Returns the meta object for the reference '{@link org.enterprisedomain.classmaker.Strategy#getState <em>State</em>}'.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @return the meta object for the reference '<em>State</em>'.
 	 * @see org.enterprisedomain.classmaker.Strategy#getState()
 	 * @see #getStrategy()
@@ -2439,10 +2289,8 @@ public interface ClassMakerPackage extends EPackage {
 	EClass getItem();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.Item#getModelName <em>Model
-	 * Name</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.Item#getModelName <em>Model Name</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute '<em>Model Name</em>'.
 	 * @see org.enterprisedomain.classmaker.Item#getModelName()
 	 * @see #getItem()
@@ -2463,10 +2311,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getItem_Phase();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.Item#getLanguage <em>Language</em>}'.
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.Item#getLanguage <em>Language</em>}'.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @return the meta object for the attribute '<em>Language</em>'.
 	 * @see org.enterprisedomain.classmaker.Item#getLanguage()
 	 * @see #getItem()
@@ -2475,12 +2321,9 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getItem_Language();
 
 	/**
-	 * Returns the meta object for the containment reference
-	 * '{@link org.enterprisedomain.classmaker.Item#getDomainModel <em>Domain
-	 * Model</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @return the meta object for the containment reference '<em>Domain
-	 *         Model</em>'.
+	 * Returns the meta object for the containment reference '{@link org.enterprisedomain.classmaker.Item#getDomainModel <em>Domain Model</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
+	 * @return the meta object for the containment reference '<em>Domain Model</em>'.
 	 * @see org.enterprisedomain.classmaker.Item#getDomainModel()
 	 * @see #getItem()
 	 * @generated
@@ -2488,10 +2331,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getItem_DomainModel();
 
 	/**
-	 * Returns the meta object for the map
-	 * '{@link org.enterprisedomain.classmaker.Item#getCustomizers
-	 * <em>Customizers</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the map '{@link org.enterprisedomain.classmaker.Item#getCustomizers <em>Customizers</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the map '<em>Customizers</em>'.
 	 * @see org.enterprisedomain.classmaker.Item#getCustomizers()
 	 * @see #getItem()
@@ -2500,10 +2341,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getItem_Customizers();
 
 	/**
-	 * Returns the meta object for the reference
-	 * '{@link org.enterprisedomain.classmaker.Item#getParent <em>Parent</em>}'.
+	 * Returns the meta object for the reference '{@link org.enterprisedomain.classmaker.Item#getParent <em>Parent</em>}'.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @return the meta object for the reference '<em>Parent</em>'.
 	 * @see org.enterprisedomain.classmaker.Item#getParent()
 	 * @see #getItem()
@@ -2512,10 +2351,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getItem_Parent();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.Item#getLocale <em>Locale</em>}'.
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.Item#getLocale <em>Locale</em>}'.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @return the meta object for the attribute '<em>Locale</em>'.
 	 * @see org.enterprisedomain.classmaker.Item#getLocale()
 	 * @see #getItem()
@@ -2524,10 +2361,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getItem_Locale();
 
 	/**
-	 * Returns the meta object for the reference
-	 * '{@link org.enterprisedomain.classmaker.Item#getProject <em>Project</em>}'.
+	 * Returns the meta object for the reference '{@link org.enterprisedomain.classmaker.Item#getProject <em>Project</em>}'.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @return the meta object for the reference '<em>Project</em>'.
 	 * @see org.enterprisedomain.classmaker.Item#getProject()
 	 * @see #getItem()
@@ -2559,12 +2394,9 @@ public interface ClassMakerPackage extends EPackage {
 	EClass getWorkspace();
 
 	/**
-	 * Returns the meta object for the containment reference list
-	 * '{@link org.enterprisedomain.classmaker.Workspace#getProjects
-	 * <em>Projects</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @return the meta object for the containment reference list
-	 *         '<em>Projects</em>'.
+	 * Returns the meta object for the containment reference list '{@link org.enterprisedomain.classmaker.Workspace#getProjects <em>Projects</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
+	 * @return the meta object for the containment reference list '<em>Projects</em>'.
 	 * @see org.enterprisedomain.classmaker.Workspace#getProjects()
 	 * @see #getWorkspace()
 	 * @generated
@@ -2572,10 +2404,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getWorkspace_Projects();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.Workspace#getResourceSet <em>Resource
-	 * Set</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.Workspace#getResourceSet <em>Resource Set</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute '<em>Resource Set</em>'.
 	 * @see org.enterprisedomain.classmaker.Workspace#getResourceSet()
 	 * @see #getWorkspace()
@@ -2584,10 +2414,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getWorkspace_ResourceSet();
 
 	/**
-	 * Returns the meta object for the map
-	 * '{@link org.enterprisedomain.classmaker.Workspace#getCustomizers
-	 * <em>Customizers</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the map '{@link org.enterprisedomain.classmaker.Workspace#getCustomizers <em>Customizers</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the map '<em>Customizers</em>'.
 	 * @see org.enterprisedomain.classmaker.Workspace#getCustomizers()
 	 * @see #getWorkspace()
@@ -2596,10 +2424,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getWorkspace_Customizers();
 
 	/**
-	 * Returns the meta object for the container reference
-	 * '{@link org.enterprisedomain.classmaker.Workspace#getService
-	 * <em>Service</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the container reference '{@link org.enterprisedomain.classmaker.Workspace#getService <em>Service</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the container reference '<em>Service</em>'.
 	 * @see org.enterprisedomain.classmaker.Workspace#getService()
 	 * @see #getWorkspace()
@@ -2608,10 +2434,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getWorkspace_Service();
 
 	/**
-	 * Returns the meta object for the reference
-	 * '{@link org.enterprisedomain.classmaker.Workspace#getSCMRegistry <em>SCM
-	 * Registry</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the reference '{@link org.enterprisedomain.classmaker.Workspace#getSCMRegistry <em>SCM Registry</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the reference '<em>SCM Registry</em>'.
 	 * @see org.enterprisedomain.classmaker.Workspace#getSCMRegistry()
 	 * @see #getWorkspace()
@@ -2620,22 +2444,19 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getWorkspace_SCMRegistry();
 
 	/**
-	 * Returns the meta object for class '{@link java.util.Map.Entry <em>Long To
-	 * State Map Entry</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for class '{@link java.util.Map.Entry <em>Long To State Map Entry</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for class '<em>Long To State Map Entry</em>'.
 	 * @see java.util.Map.Entry
 	 * @model keyDataType="org.eclipse.emf.ecore.ELongObject"
-	 *        valueType="org.enterprisedomain.classmaker.State"
-	 *        valueContainment="true"
+	 *        valueType="org.enterprisedomain.classmaker.State" valueContainment="true"
 	 * @generated
 	 */
 	EClass getLongToStateMapEntry();
 
 	/**
-	 * Returns the meta object for the attribute '{@link java.util.Map.Entry
-	 * <em>Key</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute '{@link java.util.Map.Entry <em>Key</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute '<em>Key</em>'.
 	 * @see java.util.Map.Entry
 	 * @see #getLongToStateMapEntry()
@@ -2644,10 +2465,9 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getLongToStateMapEntry_Key();
 
 	/**
-	 * Returns the meta object for the containment reference
-	 * '{@link java.util.Map.Entry <em>Value</em>}'. <!-- begin-user-doc --> <!--
+	 * Returns the meta object for the containment reference '{@link java.util.Map.Entry <em>Value</em>}'.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @return the meta object for the containment reference '<em>Value</em>'.
 	 * @see java.util.Map.Entry
 	 * @see #getLongToStateMapEntry()
@@ -2656,22 +2476,19 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getLongToStateMapEntry_Value();
 
 	/**
-	 * Returns the meta object for class '{@link java.util.Map.Entry <em>Version To
-	 * Revision Map Entry</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for class '{@link java.util.Map.Entry <em>Version To Revision Map Entry</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for class '<em>Version To Revision Map Entry</em>'.
 	 * @see java.util.Map.Entry
 	 * @model keyDataType="org.enterprisedomain.classmaker.OSGiVersion"
-	 *        valueType="org.enterprisedomain.classmaker.Revision"
-	 *        valueContainment="true"
+	 *        valueType="org.enterprisedomain.classmaker.Revision" valueContainment="true"
 	 * @generated
 	 */
 	EClass getVersionToRevisionMapEntry();
 
 	/**
-	 * Returns the meta object for the attribute '{@link java.util.Map.Entry
-	 * <em>Key</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute '{@link java.util.Map.Entry <em>Key</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute '<em>Key</em>'.
 	 * @see java.util.Map.Entry
 	 * @see #getVersionToRevisionMapEntry()
@@ -2680,10 +2497,9 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getVersionToRevisionMapEntry_Key();
 
 	/**
-	 * Returns the meta object for the containment reference
-	 * '{@link java.util.Map.Entry <em>Value</em>}'. <!-- begin-user-doc --> <!--
+	 * Returns the meta object for the containment reference '{@link java.util.Map.Entry <em>Value</em>}'.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @return the meta object for the containment reference '<em>Value</em>'.
 	 * @see java.util.Map.Entry
 	 * @see #getVersionToRevisionMapEntry()
@@ -2692,10 +2508,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getVersionToRevisionMapEntry_Value();
 
 	/**
-	 * Returns the meta object for class
-	 * '{@link org.eclipse.core.runtime.jobs.ISchedulingRule <em>IScheduling
-	 * Rule</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for class '{@link org.eclipse.core.runtime.jobs.ISchedulingRule <em>IScheduling Rule</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for class '<em>IScheduling Rule</em>'.
 	 * @see org.eclipse.core.runtime.jobs.ISchedulingRule
 	 * @model instanceClass="org.eclipse.core.runtime.jobs.ISchedulingRule"
@@ -2704,9 +2518,8 @@ public interface ClassMakerPackage extends EPackage {
 	EClass getISchedulingRule();
 
 	/**
-	 * Returns the meta object for class '{@link java.util.concurrent.Executor
-	 * <em>Executor</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for class '{@link java.util.concurrent.Executor <em>Executor</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for class '<em>Executor</em>'.
 	 * @see java.util.concurrent.Executor
 	 * @model instanceClass="java.util.concurrent.Executor"
@@ -2715,9 +2528,8 @@ public interface ClassMakerPackage extends EPackage {
 	EClass getExecutor();
 
 	/**
-	 * Returns the meta object for class '{@link java.util.concurrent.Future
-	 * <em>Future</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for class '{@link java.util.concurrent.Future <em>Future</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for class '<em>Future</em>'.
 	 * @see java.util.concurrent.Future
 	 * @model instanceClass="java.util.concurrent.Future" typeParameters="V"
@@ -2726,10 +2538,8 @@ public interface ClassMakerPackage extends EPackage {
 	EClass getFuture();
 
 	/**
-	 * Returns the meta object for class
-	 * '{@link org.eclipse.core.runtime.IAdapterFactory <em>IAdapter Factory</em>}'.
+	 * Returns the meta object for class '{@link org.eclipse.core.runtime.IAdapterFactory <em>IAdapter Factory</em>}'.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @return the meta object for class '<em>IAdapter Factory</em>'.
 	 * @see org.eclipse.core.runtime.IAdapterFactory
 	 * @model instanceClass="org.eclipse.core.runtime.IAdapterFactory"
@@ -2738,10 +2548,8 @@ public interface ClassMakerPackage extends EPackage {
 	EClass getIAdapterFactory();
 
 	/**
-	 * Returns the meta object for class
-	 * '{@link org.enterprisedomain.classmaker.Customizer <em>Customizer</em>}'.
+	 * Returns the meta object for class '{@link org.enterprisedomain.classmaker.Customizer <em>Customizer</em>}'.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @return the meta object for class '<em>Customizer</em>'.
 	 * @see org.enterprisedomain.classmaker.Customizer
 	 * @generated
@@ -2749,10 +2557,8 @@ public interface ClassMakerPackage extends EPackage {
 	EClass getCustomizer();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.Customizer#getRank <em>Rank</em>}'.
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.Customizer#getRank <em>Rank</em>}'.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @return the meta object for the attribute '<em>Rank</em>'.
 	 * @see org.enterprisedomain.classmaker.Customizer#getRank()
 	 * @see #getCustomizer()
@@ -2761,10 +2567,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getCustomizer_Rank();
 
 	/**
-	 * Returns the meta object for class
-	 * '{@link org.enterprisedomain.classmaker.StageQualifier <em>Stage
-	 * Qualifier</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for class '{@link org.enterprisedomain.classmaker.StageQualifier <em>Stage Qualifier</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for class '<em>Stage Qualifier</em>'.
 	 * @see org.enterprisedomain.classmaker.StageQualifier
 	 * @generated
@@ -2772,10 +2576,8 @@ public interface ClassMakerPackage extends EPackage {
 	EClass getStageQualifier();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.StageQualifier#getStage
-	 * <em>Stage</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.StageQualifier#getStage <em>Stage</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute '<em>Stage</em>'.
 	 * @see org.enterprisedomain.classmaker.StageQualifier#getStage()
 	 * @see #getStageQualifier()
@@ -2784,10 +2586,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getStageQualifier_Stage();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.StageQualifier#getStep
-	 * <em>Step</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.StageQualifier#getStep <em>Step</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute '<em>Step</em>'.
 	 * @see org.enterprisedomain.classmaker.StageQualifier#getStep()
 	 * @see #getStageQualifier()
@@ -2796,24 +2596,20 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getStageQualifier_Step();
 
 	/**
-	 * Returns the meta object for class '{@link java.util.Map.Entry <em>Stage
-	 * Qualifier To Customizer Map Entry</em>}'. <!-- begin-user-doc --> <!--
+	 * Returns the meta object for class '{@link java.util.Map.Entry <em>Stage Qualifier To Customizer Map Entry</em>}'.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
-	 * @return the meta object for class '<em>Stage Qualifier To Customizer Map
-	 *         Entry</em>'.
+	 * @return the meta object for class '<em>Stage Qualifier To Customizer Map Entry</em>'.
 	 * @see java.util.Map.Entry
 	 * @model keyType="org.enterprisedomain.classmaker.StageQualifier"
-	 *        valueType="org.enterprisedomain.classmaker.Customizer"
-	 *        valueContainment="true"
+	 *        valueType="org.enterprisedomain.classmaker.Customizer" valueContainment="true"
 	 * @generated
 	 */
 	EClass getStageQualifierToCustomizerMapEntry();
 
 	/**
-	 * Returns the meta object for the reference '{@link java.util.Map.Entry
-	 * <em>Key</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the reference '{@link java.util.Map.Entry <em>Key</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the reference '<em>Key</em>'.
 	 * @see java.util.Map.Entry
 	 * @see #getStageQualifierToCustomizerMapEntry()
@@ -2822,10 +2618,9 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getStageQualifierToCustomizerMapEntry_Key();
 
 	/**
-	 * Returns the meta object for the containment reference
-	 * '{@link java.util.Map.Entry <em>Value</em>}'. <!-- begin-user-doc --> <!--
+	 * Returns the meta object for the containment reference '{@link java.util.Map.Entry <em>Value</em>}'.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @return the meta object for the containment reference '<em>Value</em>'.
 	 * @see java.util.Map.Entry
 	 * @see #getStageQualifierToCustomizerMapEntry()
@@ -2857,10 +2652,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getProject_Name();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.Project#getProjectName <em>Project
-	 * Name</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.Project#getProjectName <em>Project Name</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute '<em>Project Name</em>'.
 	 * @see org.enterprisedomain.classmaker.Project#getProjectName()
 	 * @see #getProject()
@@ -2869,10 +2662,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getProject_ProjectName();
 
 	/**
-	 * Returns the meta object for the attribute list
-	 * '{@link org.enterprisedomain.classmaker.Project#getChildren
-	 * <em>Children</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute list '{@link org.enterprisedomain.classmaker.Project#getChildren <em>Children</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute list '<em>Children</em>'.
 	 * @see org.enterprisedomain.classmaker.Project#getChildren()
 	 * @see #getProject()
@@ -2881,10 +2672,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getProject_Children();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.Project#isDirty <em>Dirty</em>}'.
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.Project#isDirty <em>Dirty</em>}'.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @return the meta object for the attribute '<em>Dirty</em>'.
 	 * @see org.enterprisedomain.classmaker.Project#isDirty()
 	 * @see #getProject()
@@ -2893,10 +2682,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getProject_Dirty();
 
 	/**
-	 * Returns the meta object for the container reference
-	 * '{@link org.enterprisedomain.classmaker.Project#getWorkspace
-	 * <em>Workspace</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the container reference '{@link org.enterprisedomain.classmaker.Project#getWorkspace <em>Workspace</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the container reference '<em>Workspace</em>'.
 	 * @see org.enterprisedomain.classmaker.Project#getWorkspace()
 	 * @see #getProject()
@@ -2905,10 +2692,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getProject_Workspace();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.Project#getResourcePath <em>Resource
-	 * Path</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.Project#getResourcePath <em>Resource Path</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute '<em>Resource Path</em>'.
 	 * @see org.enterprisedomain.classmaker.Project#getResourcePath()
 	 * @see #getProject()
@@ -2917,13 +2702,10 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getProject_ResourcePath();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.Project#isNeedCompletionNotification
-	 * <em>Need Completion Notification</em>}'. <!-- begin-user-doc --> <!--
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.Project#isNeedCompletionNotification <em>Need Completion Notification</em>}'.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
-	 * @return the meta object for the attribute '<em>Need Completion
-	 *         Notification</em>'.
+	 * @return the meta object for the attribute '<em>Need Completion Notification</em>'.
 	 * @see org.enterprisedomain.classmaker.Project#isNeedCompletionNotification()
 	 * @see #getProject()
 	 * @generated
@@ -2931,13 +2713,10 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getProject_NeedCompletionNotification();
 
 	/**
-	 * Returns the meta object for the reference
-	 * '{@link org.enterprisedomain.classmaker.Project#getCompletionNotificationAdapter
-	 * <em>Completion Notification Adapter</em>}'. <!-- begin-user-doc --> <!--
+	 * Returns the meta object for the reference '{@link org.enterprisedomain.classmaker.Project#getCompletionNotificationAdapter <em>Completion Notification Adapter</em>}'.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
-	 * @return the meta object for the reference '<em>Completion Notification
-	 *         Adapter</em>'.
+	 * @return the meta object for the reference '<em>Completion Notification Adapter</em>'.
 	 * @see org.enterprisedomain.classmaker.Project#getCompletionNotificationAdapter()
 	 * @see #getProject()
 	 * @generated
@@ -2945,13 +2724,10 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getProject_CompletionNotificationAdapter();
 
 	/**
-	 * Returns the meta object for the reference
-	 * '{@link org.enterprisedomain.classmaker.Project#getResourceReloadListener
-	 * <em>Resource Reload Listener</em>}'. <!-- begin-user-doc --> <!--
+	 * Returns the meta object for the reference '{@link org.enterprisedomain.classmaker.Project#getResourceReloadListener <em>Resource Reload Listener</em>}'.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
-	 * @return the meta object for the reference '<em>Resource Reload
-	 *         Listener</em>'.
+	 * @return the meta object for the reference '<em>Resource Reload Listener</em>'.
 	 * @see org.enterprisedomain.classmaker.Project#getResourceReloadListener()
 	 * @see #getProject()
 	 * @generated
@@ -2959,10 +2735,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getProject_ResourceReloadListener();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.Project#isSavingResource <em>Saving
-	 * Resource</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.Project#isSavingResource <em>Saving Resource</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute '<em>Saving Resource</em>'.
 	 * @see org.enterprisedomain.classmaker.Project#isSavingResource()
 	 * @see #getProject()
@@ -2971,10 +2745,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getProject_SavingResource();
 
 	/**
-	 * Returns the meta object for the reference
-	 * '{@link org.enterprisedomain.classmaker.Project#getRevision
-	 * <em>Revision</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the reference '{@link org.enterprisedomain.classmaker.Project#getRevision <em>Revision</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the reference '<em>Revision</em>'.
 	 * @see org.enterprisedomain.classmaker.Project#getRevision()
 	 * @see #getProject()
@@ -2983,10 +2755,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getProject_Revision();
 
 	/**
-	 * Returns the meta object for the map
-	 * '{@link org.enterprisedomain.classmaker.Project#getRevisions
-	 * <em>Revisions</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the map '{@link org.enterprisedomain.classmaker.Project#getRevisions <em>Revisions</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the map '<em>Revisions</em>'.
 	 * @see org.enterprisedomain.classmaker.Project#getRevisions()
 	 * @see #getProject()
@@ -2995,10 +2765,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getProject_Revisions();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.Project#getProjectVersion <em>Project
-	 * Version</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.Project#getProjectVersion <em>Project Version</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute '<em>Project Version</em>'.
 	 * @see org.enterprisedomain.classmaker.Project#getProjectVersion()
 	 * @see #getProject()
@@ -3020,10 +2788,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getProject_SelectRevealHandler();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.Project#getVersion
-	 * <em>Version</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.Project#getVersion <em>Version</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute '<em>Version</em>'.
 	 * @see org.enterprisedomain.classmaker.Project#getVersion()
 	 * @see #getProject()
@@ -3032,10 +2798,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getProject_Version();
 
 	/**
-	 * Returns the meta object for the reference
-	 * '{@link org.enterprisedomain.classmaker.Project#getState <em>State</em>}'.
+	 * Returns the meta object for the reference '{@link org.enterprisedomain.classmaker.Project#getState <em>State</em>}'.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @return the meta object for the reference '<em>State</em>'.
 	 * @see org.enterprisedomain.classmaker.Project#getState()
 	 * @see #getProject()
@@ -3069,10 +2833,8 @@ public interface ClassMakerPackage extends EPackage {
 	EClass getModelPair();
 
 	/**
-	 * Returns the meta object for the containment reference
-	 * '{@link org.enterprisedomain.classmaker.ModelPair#getDynamic
-	 * <em>Dynamic</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the containment reference '{@link org.enterprisedomain.classmaker.ModelPair#getDynamic <em>Dynamic</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the containment reference '<em>Dynamic</em>'.
 	 * @see org.enterprisedomain.classmaker.ModelPair#getDynamic()
 	 * @see #getModelPair()
@@ -3081,10 +2843,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getModelPair_Dynamic();
 
 	/**
-	 * Returns the meta object for the reference
-	 * '{@link org.enterprisedomain.classmaker.ModelPair#getGenerated
-	 * <em>Generated</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the reference '{@link org.enterprisedomain.classmaker.ModelPair#getGenerated <em>Generated</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the reference '<em>Generated</em>'.
 	 * @see org.enterprisedomain.classmaker.ModelPair#getGenerated()
 	 * @see #getModelPair()
@@ -3093,10 +2853,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getModelPair_Generated();
 
 	/**
-	 * Returns the meta object for the container reference
-	 * '{@link org.enterprisedomain.classmaker.ModelPair#getParent
-	 * <em>Parent</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the container reference '{@link org.enterprisedomain.classmaker.ModelPair#getParent <em>Parent</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the container reference '<em>Parent</em>'.
 	 * @see org.enterprisedomain.classmaker.ModelPair#getParent()
 	 * @see #getModelPair()
@@ -3105,10 +2863,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getModelPair_Parent();
 
 	/**
-	 * Returns the meta object for class
-	 * '{@link org.enterprisedomain.classmaker.SCMOperator <em>SCM Operator</em>}'.
+	 * Returns the meta object for class '{@link org.enterprisedomain.classmaker.SCMOperator <em>SCM Operator</em>}'.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @return the meta object for class '<em>SCM Operator</em>'.
 	 * @see org.enterprisedomain.classmaker.SCMOperator
 	 * @generated
@@ -3116,10 +2872,8 @@ public interface ClassMakerPackage extends EPackage {
 	EClass getSCMOperator();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.SCMOperator#getProjectName
-	 * <em>Project Name</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.SCMOperator#getProjectName <em>Project Name</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute '<em>Project Name</em>'.
 	 * @see org.enterprisedomain.classmaker.SCMOperator#getProjectName()
 	 * @see #getSCMOperator()
@@ -3128,10 +2882,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getSCMOperator_ProjectName();
 
 	/**
-	 * Returns the meta object for the reference
-	 * '{@link org.enterprisedomain.classmaker.SCMOperator#getRegistry
-	 * <em>Registry</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the reference '{@link org.enterprisedomain.classmaker.SCMOperator#getRegistry <em>Registry</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the reference '<em>Registry</em>'.
 	 * @see org.enterprisedomain.classmaker.SCMOperator#getRegistry()
 	 * @see #getSCMOperator()
@@ -3140,10 +2892,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getSCMOperator_Registry();
 
 	/**
-	 * Returns the meta object for class
-	 * '{@link org.enterprisedomain.classmaker.SCMRegistry <em>SCM Registry</em>}'.
+	 * Returns the meta object for class '{@link org.enterprisedomain.classmaker.SCMRegistry <em>SCM Registry</em>}'.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @return the meta object for class '<em>SCM Registry</em>'.
 	 * @see org.enterprisedomain.classmaker.SCMRegistry
 	 * @generated
@@ -3163,10 +2913,8 @@ public interface ClassMakerPackage extends EPackage {
 	EClass getResource();
 
 	/**
-	 * Returns the meta object for class
-	 * '{@link org.enterprisedomain.classmaker.ResourceChangeListener <em>Resource
-	 * Change Listener</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for class '{@link org.enterprisedomain.classmaker.ResourceChangeListener <em>Resource Change Listener</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for class '<em>Resource Change Listener</em>'.
 	 * @see org.enterprisedomain.classmaker.ResourceChangeListener
 	 * @generated
@@ -3174,10 +2922,8 @@ public interface ClassMakerPackage extends EPackage {
 	EClass getResourceChangeListener();
 
 	/**
-	 * Returns the meta object for class
-	 * '{@link org.eclipse.emf.common.notify.Notification <em>Notification</em>}'.
+	 * Returns the meta object for class '{@link org.eclipse.emf.common.notify.Notification <em>Notification</em>}'.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @return the meta object for class '<em>Notification</em>'.
 	 * @see org.eclipse.emf.common.notify.Notification
 	 * @model instanceClass="org.eclipse.emf.common.notify.Notification"
@@ -3186,10 +2932,8 @@ public interface ClassMakerPackage extends EPackage {
 	EClass getNotification();
 
 	/**
-	 * Returns the meta object for class
-	 * '{@link org.enterprisedomain.classmaker.CompletionListener <em>Completion
-	 * Listener</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for class '{@link org.enterprisedomain.classmaker.CompletionListener <em>Completion Listener</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for class '<em>Completion Listener</em>'.
 	 * @see org.enterprisedomain.classmaker.CompletionListener
 	 * @generated
@@ -3209,10 +2953,8 @@ public interface ClassMakerPackage extends EPackage {
 	EClass getWorker();
 
 	/**
-	 * Returns the meta object for class
-	 * '{@link org.enterprisedomain.classmaker.SelectRevealHandler <em>Select Reveal
-	 * Handler</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for class '{@link org.enterprisedomain.classmaker.SelectRevealHandler <em>Select Reveal Handler</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for class '<em>Select Reveal Handler</em>'.
 	 * @see org.enterprisedomain.classmaker.SelectRevealHandler
 	 * @generated
@@ -3220,10 +2962,8 @@ public interface ClassMakerPackage extends EPackage {
 	EClass getSelectRevealHandler();
 
 	/**
-	 * Returns the meta object for class
-	 * '{@link org.enterprisedomain.classmaker.ResourceAdapter <em>Resource
-	 * Adapter</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for class '{@link org.enterprisedomain.classmaker.ResourceAdapter <em>Resource Adapter</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for class '<em>Resource Adapter</em>'.
 	 * @see org.enterprisedomain.classmaker.ResourceAdapter
 	 * @generated
@@ -3231,10 +2971,8 @@ public interface ClassMakerPackage extends EPackage {
 	EClass getResourceAdapter();
 
 	/**
-	 * Returns the meta object for the reference
-	 * '{@link org.enterprisedomain.classmaker.ResourceAdapter#getResource
-	 * <em>Resource</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the reference '{@link org.enterprisedomain.classmaker.ResourceAdapter#getResource <em>Resource</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the reference '<em>Resource</em>'.
 	 * @see org.enterprisedomain.classmaker.ResourceAdapter#getResource()
 	 * @see #getResourceAdapter()
@@ -3243,10 +2981,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getResourceAdapter_Resource();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.ResourceAdapter#getFilename
-	 * <em>Filename</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.ResourceAdapter#getFilename <em>Filename</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute '<em>Filename</em>'.
 	 * @see org.enterprisedomain.classmaker.ResourceAdapter#getFilename()
 	 * @see #getResourceAdapter()
@@ -3255,10 +2991,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getResourceAdapter_Filename();
 
 	/**
-	 * Returns the meta object for the container reference
-	 * '{@link org.enterprisedomain.classmaker.ResourceAdapter#getProject
-	 * <em>Project</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the container reference '{@link org.enterprisedomain.classmaker.ResourceAdapter#getProject <em>Project</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the container reference '<em>Project</em>'.
 	 * @see org.enterprisedomain.classmaker.ResourceAdapter#getProject()
 	 * @see #getResourceAdapter()
@@ -3267,10 +3001,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getResourceAdapter_Project();
 
 	/**
-	 * Returns the meta object for class
-	 * '{@link org.enterprisedomain.classmaker.ClassMakerService <em>Service</em>}'.
+	 * Returns the meta object for class '{@link org.enterprisedomain.classmaker.ClassMakerService <em>Service</em>}'.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @return the meta object for class '<em>Service</em>'.
 	 * @see org.enterprisedomain.classmaker.ClassMakerService
 	 * @generated
@@ -3278,10 +3010,8 @@ public interface ClassMakerPackage extends EPackage {
 	EClass getClassMakerService();
 
 	/**
-	 * Returns the meta object for the containment reference
-	 * '{@link org.enterprisedomain.classmaker.ClassMakerService#getWorkspace
-	 * <em>Workspace</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the containment reference '{@link org.enterprisedomain.classmaker.ClassMakerService#getWorkspace <em>Workspace</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the containment reference '<em>Workspace</em>'.
 	 * @see org.enterprisedomain.classmaker.ClassMakerService#getWorkspace()
 	 * @see #getClassMakerService()
@@ -3301,10 +3031,8 @@ public interface ClassMakerPackage extends EPackage {
 	EClass getBlueprint();
 
 	/**
-	 * Returns the meta object for the reference
-	 * '{@link org.enterprisedomain.classmaker.Blueprint#getDynamicModel <em>Dynamic
-	 * Model</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the reference '{@link org.enterprisedomain.classmaker.Blueprint#getDynamicModel <em>Dynamic Model</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the reference '<em>Dynamic Model</em>'.
 	 * @see org.enterprisedomain.classmaker.Blueprint#getDynamicModel()
 	 * @see #getBlueprint()
@@ -3313,10 +3041,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getBlueprint_DynamicModel();
 
 	/**
-	 * Returns the meta object for the attribute list
-	 * '{@link org.enterprisedomain.classmaker.Blueprint#getDependencies
-	 * <em>Dependencies</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute list '{@link org.enterprisedomain.classmaker.Blueprint#getDependencies <em>Dependencies</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute list '<em>Dependencies</em>'.
 	 * @see org.enterprisedomain.classmaker.Blueprint#getDependencies()
 	 * @see #getBlueprint()
@@ -3339,10 +3065,8 @@ public interface ClassMakerPackage extends EPackage {
 	EReference getBlueprint_CompletionListeners();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.Blueprint#isEdit <em>Edit</em>}'.
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.Blueprint#isEdit <em>Edit</em>}'.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @return the meta object for the attribute '<em>Edit</em>'.
 	 * @see org.enterprisedomain.classmaker.Blueprint#isEdit()
 	 * @see #getBlueprint()
@@ -3351,10 +3075,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getBlueprint_Edit();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.Blueprint#isEditor <em>Editor</em>}'.
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.Blueprint#isEditor <em>Editor</em>}'.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @return the meta object for the attribute '<em>Editor</em>'.
 	 * @see org.enterprisedomain.classmaker.Blueprint#isEditor()
 	 * @see #getBlueprint()
@@ -3363,11 +3085,9 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getBlueprint_Editor();
 
 	/**
-	 * Returns the meta object for class
-	 * '{@link org.enterprisedomain.classmaker.CompletionNotificationAdapter
-	 * <em>Completion Notification Adapter</em>}'. <!-- begin-user-doc --> <!--
+	 * Returns the meta object for class '{@link org.enterprisedomain.classmaker.CompletionNotificationAdapter <em>Completion Notification Adapter</em>}'.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @return the meta object for class '<em>Completion Notification Adapter</em>'.
 	 * @see org.enterprisedomain.classmaker.CompletionNotificationAdapter
 	 * @generated
@@ -3375,10 +3095,8 @@ public interface ClassMakerPackage extends EPackage {
 	EClass getCompletionNotificationAdapter();
 
 	/**
-	 * Returns the meta object for the attribute
-	 * '{@link org.enterprisedomain.classmaker.CompletionNotificationAdapter#getError
-	 * <em>Error</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the attribute '{@link org.enterprisedomain.classmaker.CompletionNotificationAdapter#getError <em>Error</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the attribute '<em>Error</em>'.
 	 * @see org.enterprisedomain.classmaker.CompletionNotificationAdapter#getError()
 	 * @see #getCompletionNotificationAdapter()
@@ -3387,10 +3105,8 @@ public interface ClassMakerPackage extends EPackage {
 	EAttribute getCompletionNotificationAdapter_Error();
 
 	/**
-	 * Returns the meta object for the reference
-	 * '{@link org.enterprisedomain.classmaker.CompletionNotificationAdapter#getProject
-	 * <em>Project</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for the reference '{@link org.enterprisedomain.classmaker.CompletionNotificationAdapter#getProject <em>Project</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for the reference '<em>Project</em>'.
 	 * @see org.enterprisedomain.classmaker.CompletionNotificationAdapter#getProject()
 	 * @see #getCompletionNotificationAdapter()
@@ -3529,9 +3245,8 @@ public interface ClassMakerPackage extends EPackage {
 	EDataType getURI();
 
 	/**
-	 * Returns the meta object for data type '{@link java.lang.String
-	 * <em>Name</em>}'. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the meta object for data type '{@link java.lang.String <em>Name</em>}'.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @return the meta object for data type '<em>Name</em>'.
 	 * @see java.lang.String
 	 * @model instanceClass="java.lang.String"
@@ -3571,15 +3286,12 @@ public interface ClassMakerPackage extends EPackage {
 	 * <li>and each data type</li>
 	 * </ul>
 	 * <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	interface Literals {
 		/**
-		 * The meta object literal for the
-		 * '{@link org.enterprisedomain.classmaker.impl.ContributionImpl
-		 * <em>Contribution</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '{@link org.enterprisedomain.classmaker.impl.ContributionImpl <em>Contribution</em>}' class.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @see org.enterprisedomain.classmaker.impl.ContributionImpl
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getContribution()
 		 * @generated
@@ -3587,26 +3299,22 @@ public interface ClassMakerPackage extends EPackage {
 		EClass CONTRIBUTION = eINSTANCE.getContribution();
 
 		/**
-		 * The meta object literal for the '<em><b>Dependencies</b></em>' attribute list
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Dependencies</b></em>' attribute list feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EAttribute CONTRIBUTION__DEPENDENCIES = eINSTANCE.getContribution_Dependencies();
 
 		/**
-		 * The meta object literal for the '<em><b>Latest Version</b></em>' attribute
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Latest Version</b></em>' attribute feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EAttribute CONTRIBUTION__LATEST_VERSION = eINSTANCE.getContribution_LatestVersion();
 
 		/**
-		 * The meta object literal for the
-		 * '{@link org.enterprisedomain.classmaker.impl.RevisionImpl <em>Revision</em>}'
-		 * class. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '{@link org.enterprisedomain.classmaker.impl.RevisionImpl <em>Revision</em>}' class.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @see org.enterprisedomain.classmaker.impl.RevisionImpl
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getRevision()
 		 * @generated
@@ -3616,15 +3324,13 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>State</b></em>' reference feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EReference REVISION__STATE = eINSTANCE.getRevision_State();
 
 		/**
-		 * The meta object literal for the '<em><b>Timestamp</b></em>' attribute
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Timestamp</b></em>' attribute feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EAttribute REVISION__TIMESTAMP = eINSTANCE.getRevision_Timestamp();
@@ -3632,15 +3338,13 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>State History</b></em>' map feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EReference REVISION__STATE_HISTORY = eINSTANCE.getRevision_StateHistory();
 
 		/**
-		 * The meta object literal for the '<em><b>Latest Timestamp</b></em>' attribute
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Latest Timestamp</b></em>' attribute feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EAttribute REVISION__LATEST_TIMESTAMP = eINSTANCE.getRevision_LatestTimestamp();
@@ -3648,16 +3352,13 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Version</b></em>' attribute feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EAttribute REVISION__VERSION = eINSTANCE.getRevision_Version();
 
 		/**
-		 * The meta object literal for the
-		 * '{@link org.enterprisedomain.classmaker.impl.StateImpl <em>State</em>}'
-		 * class. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '{@link org.enterprisedomain.classmaker.impl.StateImpl <em>State</em>}' class.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @see org.enterprisedomain.classmaker.impl.StateImpl
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getState()
 		 * @generated
@@ -3665,33 +3366,29 @@ public interface ClassMakerPackage extends EPackage {
 		EClass STATE = eINSTANCE.getState();
 
 		/**
-		 * The meta object literal for the '<em><b>Package Class Name</b></em>'
-		 * attribute feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Package Class Name</b></em>' attribute feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EAttribute STATE__PACKAGE_CLASS_NAME = eINSTANCE.getState_PackageClassName();
 
 		/**
-		 * The meta object literal for the '<em><b>Edit Plugin Class Name</b></em>'
-		 * attribute feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Edit Plugin Class Name</b></em>' attribute feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EAttribute STATE__EDIT_PLUGIN_CLASS_NAME = eINSTANCE.getState_EditPluginClassName();
 
 		/**
-		 * The meta object literal for the '<em><b>Editor Plugin Class Name</b></em>'
-		 * attribute feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Editor Plugin Class Name</b></em>' attribute feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EAttribute STATE__EDITOR_PLUGIN_CLASS_NAME = eINSTANCE.getState_EditorPluginClassName();
 
 		/**
-		 * The meta object literal for the '<em><b>Required Plugins</b></em>' attribute
-		 * list feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Required Plugins</b></em>' attribute list feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EAttribute STATE__REQUIRED_PLUGINS = eINSTANCE.getState_RequiredPlugins();
@@ -3699,47 +3396,41 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Revision</b></em>' reference feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EReference STATE__REVISION = eINSTANCE.getState_Revision();
 
 		/**
-		 * The meta object literal for the '<em><b>Timestamp</b></em>' attribute
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Timestamp</b></em>' attribute feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EAttribute STATE__TIMESTAMP = eINSTANCE.getState_Timestamp();
 
 		/**
-		 * The meta object literal for the '<em><b>Deployable Unit Name</b></em>'
-		 * attribute feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Deployable Unit Name</b></em>' attribute feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EAttribute STATE__DEPLOYABLE_UNIT_NAME = eINSTANCE.getState_DeployableUnitName();
 
 		/**
-		 * The meta object literal for the '<em><b>Edit Deployable Unit Name</b></em>'
-		 * attribute feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Edit Deployable Unit Name</b></em>' attribute feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EAttribute STATE__EDIT_DEPLOYABLE_UNIT_NAME = eINSTANCE.getState_EditDeployableUnitName();
 
 		/**
-		 * The meta object literal for the '<em><b>Editor Deployable Unit Name</b></em>'
-		 * attribute feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Editor Deployable Unit Name</b></em>' attribute feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EAttribute STATE__EDITOR_DEPLOYABLE_UNIT_NAME = eINSTANCE.getState_EditorDeployableUnitName();
 
 		/**
-		 * The meta object literal for the '<em><b>Job Family</b></em>' attribute
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Job Family</b></em>' attribute feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EAttribute STATE__JOB_FAMILY = eINSTANCE.getState_JobFamily();
@@ -3747,39 +3438,34 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Resource</b></em>' reference feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EReference STATE__RESOURCE = eINSTANCE.getState_Resource();
 
 		/**
-		 * The meta object literal for the '<em><b>Commit Ids</b></em>' attribute list
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Commit Ids</b></em>' attribute list feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EAttribute STATE__COMMIT_IDS = eINSTANCE.getState_CommitIds();
 
 		/**
-		 * The meta object literal for the '<em><b>Commit Id</b></em>' attribute
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Commit Id</b></em>' attribute feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EAttribute STATE__COMMIT_ID = eINSTANCE.getState_CommitId();
 
 		/**
-		 * The meta object literal for the '<em><b>State Customizers</b></em>' map
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>State Customizers</b></em>' map feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EReference STATE__STATE_CUSTOMIZERS = eINSTANCE.getState_StateCustomizers();
 
 		/**
-		 * The meta object literal for the '<em><b>Project Name</b></em>' attribute
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Project Name</b></em>' attribute feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EAttribute STATE__PROJECT_NAME = eINSTANCE.getState_ProjectName();
@@ -3787,7 +3473,6 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Making</b></em>' attribute feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EAttribute STATE__MAKING = eINSTANCE.getState_Making();
@@ -3795,7 +3480,6 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Edit</b></em>' attribute feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EAttribute STATE__EDIT = eINSTANCE.getState_Edit();
@@ -3803,7 +3487,6 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Editor</b></em>' attribute feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EAttribute STATE__EDITOR = eINSTANCE.getState_Editor();
@@ -3811,16 +3494,13 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Strategy</b></em>' reference feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EReference STATE__STRATEGY = eINSTANCE.getState_Strategy();
 
 		/**
-		 * The meta object literal for the
-		 * '{@link org.enterprisedomain.classmaker.impl.StrategyImpl <em>Strategy</em>}'
-		 * class. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '{@link org.enterprisedomain.classmaker.impl.StrategyImpl <em>Strategy</em>}' class.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @see org.enterprisedomain.classmaker.impl.StrategyImpl
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getStrategy()
 		 * @generated
@@ -3828,33 +3508,29 @@ public interface ClassMakerPackage extends EPackage {
 		EClass STRATEGY = eINSTANCE.getStrategy();
 
 		/**
-		 * The meta object literal for the '<em><b>Generators</b></em>' reference list
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Generators</b></em>' reference list feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EReference STRATEGY__GENERATORS = eINSTANCE.getStrategy_Generators();
 
 		/**
-		 * The meta object literal for the '<em><b>Exporters</b></em>' reference list
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Exporters</b></em>' reference list feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EReference STRATEGY__EXPORTERS = eINSTANCE.getStrategy_Exporters();
 
 		/**
-		 * The meta object literal for the '<em><b>Installers</b></em>' reference list
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Installers</b></em>' reference list feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EReference STRATEGY__INSTALLERS = eINSTANCE.getStrategy_Installers();
 
 		/**
-		 * The meta object literal for the '<em><b>Loaders</b></em>' reference list
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Loaders</b></em>' reference list feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EReference STRATEGY__LOADERS = eINSTANCE.getStrategy_Loaders();
@@ -3862,16 +3538,13 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>State</b></em>' reference feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EReference STRATEGY__STATE = eINSTANCE.getStrategy_State();
 
 		/**
-		 * The meta object literal for the
-		 * '{@link org.enterprisedomain.classmaker.impl.ItemImpl <em>Item</em>}' class.
+		 * The meta object literal for the '{@link org.enterprisedomain.classmaker.impl.ItemImpl <em>Item</em>}' class.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @see org.enterprisedomain.classmaker.impl.ItemImpl
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getItem()
 		 * @generated
@@ -3879,9 +3552,8 @@ public interface ClassMakerPackage extends EPackage {
 		EClass ITEM = eINSTANCE.getItem();
 
 		/**
-		 * The meta object literal for the '<em><b>Model Name</b></em>' attribute
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Model Name</b></em>' attribute feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EAttribute ITEM__MODEL_NAME = eINSTANCE.getItem_ModelName();
@@ -3889,7 +3561,6 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Phase</b></em>' attribute feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EAttribute ITEM__PHASE = eINSTANCE.getItem_Phase();
@@ -3897,15 +3568,13 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Language</b></em>' attribute feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EAttribute ITEM__LANGUAGE = eINSTANCE.getItem_Language();
 
 		/**
-		 * The meta object literal for the '<em><b>Domain Model</b></em>' containment
-		 * reference feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Domain Model</b></em>' containment reference feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EReference ITEM__DOMAIN_MODEL = eINSTANCE.getItem_DomainModel();
@@ -3913,7 +3582,6 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Customizers</b></em>' map feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EReference ITEM__CUSTOMIZERS = eINSTANCE.getItem_Customizers();
@@ -3921,7 +3589,6 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Parent</b></em>' reference feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EReference ITEM__PARENT = eINSTANCE.getItem_Parent();
@@ -3929,7 +3596,6 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Locale</b></em>' attribute feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EAttribute ITEM__LOCALE = eINSTANCE.getItem_Locale();
@@ -3937,15 +3603,13 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Project</b></em>' reference feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EReference ITEM__PROJECT = eINSTANCE.getItem_Project();
 
 		/**
-		 * The meta object literal for the '{@link org.eclipse.emf.common.notify.Adapter
-		 * <em>Adapter</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '{@link org.eclipse.emf.common.notify.Adapter <em>Adapter</em>}' class.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @see org.eclipse.emf.common.notify.Adapter
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getAdapter()
 		 * @generated
@@ -3953,10 +3617,8 @@ public interface ClassMakerPackage extends EPackage {
 		EClass ADAPTER = eINSTANCE.getAdapter();
 
 		/**
-		 * The meta object literal for the
-		 * '{@link org.enterprisedomain.classmaker.impl.WorkspaceImpl
-		 * <em>Workspace</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '{@link org.enterprisedomain.classmaker.impl.WorkspaceImpl <em>Workspace</em>}' class.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @see org.enterprisedomain.classmaker.impl.WorkspaceImpl
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getWorkspace()
 		 * @generated
@@ -3964,17 +3626,15 @@ public interface ClassMakerPackage extends EPackage {
 		EClass WORKSPACE = eINSTANCE.getWorkspace();
 
 		/**
-		 * The meta object literal for the '<em><b>Projects</b></em>' containment
-		 * reference list feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Projects</b></em>' containment reference list feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EReference WORKSPACE__PROJECTS = eINSTANCE.getWorkspace_Projects();
 
 		/**
-		 * The meta object literal for the '<em><b>Resource Set</b></em>' attribute
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Resource Set</b></em>' attribute feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EAttribute WORKSPACE__RESOURCE_SET = eINSTANCE.getWorkspace_ResourceSet();
@@ -3982,23 +3642,20 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Customizers</b></em>' map feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EReference WORKSPACE__CUSTOMIZERS = eINSTANCE.getWorkspace_Customizers();
 
 		/**
-		 * The meta object literal for the '<em><b>Service</b></em>' container reference
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Service</b></em>' container reference feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EReference WORKSPACE__SERVICE = eINSTANCE.getWorkspace_Service();
 
 		/**
-		 * The meta object literal for the '<em><b>SCM Registry</b></em>' reference
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>SCM Registry</b></em>' reference feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EReference WORKSPACE__SCM_REGISTRY = eINSTANCE.getWorkspace_SCMRegistry();
@@ -4024,19 +3681,16 @@ public interface ClassMakerPackage extends EPackage {
 		EAttribute LONG_TO_STATE_MAP_ENTRY__KEY = eINSTANCE.getLongToStateMapEntry_Key();
 
 		/**
-		 * The meta object literal for the '<em><b>Value</b></em>' containment reference
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Value</b></em>' containment reference feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EReference LONG_TO_STATE_MAP_ENTRY__VALUE = eINSTANCE.getLongToStateMapEntry_Value();
 
 		/**
-		 * The meta object literal for the
-		 * '{@link org.enterprisedomain.classmaker.impl.VersionToRevisionMapEntryImpl
-		 * <em>Version To Revision Map Entry</em>}' class. <!-- begin-user-doc --> <!--
+		 * The meta object literal for the '{@link org.enterprisedomain.classmaker.impl.VersionToRevisionMapEntryImpl <em>Version To Revision Map Entry</em>}' class.
+		 * <!-- begin-user-doc --> <!--
 		 * end-user-doc -->
-		 * 
 		 * @see org.enterprisedomain.classmaker.impl.VersionToRevisionMapEntryImpl
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getVersionToRevisionMapEntry()
 		 * @generated
@@ -4052,18 +3706,15 @@ public interface ClassMakerPackage extends EPackage {
 		EAttribute VERSION_TO_REVISION_MAP_ENTRY__KEY = eINSTANCE.getVersionToRevisionMapEntry_Key();
 
 		/**
-		 * The meta object literal for the '<em><b>Value</b></em>' containment reference
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Value</b></em>' containment reference feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EReference VERSION_TO_REVISION_MAP_ENTRY__VALUE = eINSTANCE.getVersionToRevisionMapEntry_Value();
 
 		/**
-		 * The meta object literal for the
-		 * '{@link org.eclipse.core.runtime.jobs.ISchedulingRule <em>IScheduling
-		 * Rule</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '{@link org.eclipse.core.runtime.jobs.ISchedulingRule <em>IScheduling Rule</em>}' class.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @see org.eclipse.core.runtime.jobs.ISchedulingRule
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getISchedulingRule()
 		 * @generated
@@ -4071,9 +3722,8 @@ public interface ClassMakerPackage extends EPackage {
 		EClass ISCHEDULING_RULE = eINSTANCE.getISchedulingRule();
 
 		/**
-		 * The meta object literal for the '{@link java.util.concurrent.Executor
-		 * <em>Executor</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '{@link java.util.concurrent.Executor <em>Executor</em>}' class.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @see java.util.concurrent.Executor
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getExecutor()
 		 * @generated
@@ -4081,9 +3731,8 @@ public interface ClassMakerPackage extends EPackage {
 		EClass EXECUTOR = eINSTANCE.getExecutor();
 
 		/**
-		 * The meta object literal for the '{@link java.util.concurrent.Future
-		 * <em>Future</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '{@link java.util.concurrent.Future <em>Future</em>}' class.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @see java.util.concurrent.Future
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getFuture()
 		 * @generated
@@ -4091,10 +3740,8 @@ public interface ClassMakerPackage extends EPackage {
 		EClass FUTURE = eINSTANCE.getFuture();
 
 		/**
-		 * The meta object literal for the
-		 * '{@link org.eclipse.core.runtime.IAdapterFactory <em>IAdapter Factory</em>}'
-		 * class. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '{@link org.eclipse.core.runtime.IAdapterFactory <em>IAdapter Factory</em>}' class.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @see org.eclipse.core.runtime.IAdapterFactory
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getIAdapterFactory()
 		 * @generated
@@ -4102,10 +3749,8 @@ public interface ClassMakerPackage extends EPackage {
 		EClass IADAPTER_FACTORY = eINSTANCE.getIAdapterFactory();
 
 		/**
-		 * The meta object literal for the
-		 * '{@link org.enterprisedomain.classmaker.impl.CustomizerImpl
-		 * <em>Customizer</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '{@link org.enterprisedomain.classmaker.impl.CustomizerImpl <em>Customizer</em>}' class.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @see org.enterprisedomain.classmaker.impl.CustomizerImpl
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getCustomizer()
 		 * @generated
@@ -4115,16 +3760,13 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Rank</b></em>' attribute feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EAttribute CUSTOMIZER__RANK = eINSTANCE.getCustomizer_Rank();
 
 		/**
-		 * The meta object literal for the
-		 * '{@link org.enterprisedomain.classmaker.impl.StageQualifierImpl <em>Stage
-		 * Qualifier</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '{@link org.enterprisedomain.classmaker.impl.StageQualifierImpl <em>Stage Qualifier</em>}' class.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @see org.enterprisedomain.classmaker.impl.StageQualifierImpl
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getStageQualifier()
 		 * @generated
@@ -4134,7 +3776,6 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Stage</b></em>' attribute feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EAttribute STAGE_QUALIFIER__STAGE = eINSTANCE.getStageQualifier_Stage();
@@ -4142,17 +3783,14 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Step</b></em>' attribute feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EAttribute STAGE_QUALIFIER__STEP = eINSTANCE.getStageQualifier_Step();
 
 		/**
-		 * The meta object literal for the
-		 * '{@link org.enterprisedomain.classmaker.impl.StageQualifierToCustomizerMapEntryImpl
-		 * <em>Stage Qualifier To Customizer Map Entry</em>}' class. <!-- begin-user-doc
+		 * The meta object literal for the '{@link org.enterprisedomain.classmaker.impl.StageQualifierToCustomizerMapEntryImpl <em>Stage Qualifier To Customizer Map Entry</em>}' class.
+		 * <!-- begin-user-doc
 		 * --> <!-- end-user-doc -->
-		 * 
 		 * @see org.enterprisedomain.classmaker.impl.StageQualifierToCustomizerMapEntryImpl
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getStageQualifierToCustomizerMapEntry()
 		 * @generated
@@ -4168,19 +3806,16 @@ public interface ClassMakerPackage extends EPackage {
 		EReference STAGE_QUALIFIER_TO_CUSTOMIZER_MAP_ENTRY__KEY = eINSTANCE.getStageQualifierToCustomizerMapEntry_Key();
 
 		/**
-		 * The meta object literal for the '<em><b>Value</b></em>' containment reference
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Value</b></em>' containment reference feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EReference STAGE_QUALIFIER_TO_CUSTOMIZER_MAP_ENTRY__VALUE = eINSTANCE
 				.getStageQualifierToCustomizerMapEntry_Value();
 
 		/**
-		 * The meta object literal for the
-		 * '{@link org.enterprisedomain.classmaker.impl.ProjectImpl <em>Project</em>}'
-		 * class. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '{@link org.enterprisedomain.classmaker.impl.ProjectImpl <em>Project</em>}' class.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @see org.enterprisedomain.classmaker.impl.ProjectImpl
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getProject()
 		 * @generated
@@ -4190,23 +3825,20 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Name</b></em>' attribute feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EAttribute PROJECT__NAME = eINSTANCE.getProject_Name();
 
 		/**
-		 * The meta object literal for the '<em><b>Project Name</b></em>' attribute
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Project Name</b></em>' attribute feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EAttribute PROJECT__PROJECT_NAME = eINSTANCE.getProject_ProjectName();
 
 		/**
-		 * The meta object literal for the '<em><b>Children</b></em>' attribute list
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Children</b></em>' attribute list feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EAttribute PROJECT__CHILDREN = eINSTANCE.getProject_Children();
@@ -4214,57 +3846,50 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Dirty</b></em>' attribute feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EAttribute PROJECT__DIRTY = eINSTANCE.getProject_Dirty();
 
 		/**
-		 * The meta object literal for the '<em><b>Workspace</b></em>' container
-		 * reference feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Workspace</b></em>' container reference feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EReference PROJECT__WORKSPACE = eINSTANCE.getProject_Workspace();
 
 		/**
-		 * The meta object literal for the '<em><b>Resource Path</b></em>' attribute
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Resource Path</b></em>' attribute feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EAttribute PROJECT__RESOURCE_PATH = eINSTANCE.getProject_ResourcePath();
 
 		/**
-		 * The meta object literal for the '<em><b>Need Completion
-		 * Notification</b></em>' attribute feature. <!-- begin-user-doc --> <!--
+		 * The meta object literal for the '<em><b>Need Completion Notification</b></em>' attribute feature.
+		 * <!-- begin-user-doc --> <!--
 		 * end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EAttribute PROJECT__NEED_COMPLETION_NOTIFICATION = eINSTANCE.getProject_NeedCompletionNotification();
 
 		/**
-		 * The meta object literal for the '<em><b>Completion Notification
-		 * Adapter</b></em>' reference feature. <!-- begin-user-doc --> <!--
+		 * The meta object literal for the '<em><b>Completion Notification Adapter</b></em>' reference feature.
+		 * <!-- begin-user-doc --> <!--
 		 * end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EReference PROJECT__COMPLETION_NOTIFICATION_ADAPTER = eINSTANCE.getProject_CompletionNotificationAdapter();
 
 		/**
-		 * The meta object literal for the '<em><b>Resource Reload Listener</b></em>'
-		 * reference feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Resource Reload Listener</b></em>' reference feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EReference PROJECT__RESOURCE_RELOAD_LISTENER = eINSTANCE.getProject_ResourceReloadListener();
 
 		/**
-		 * The meta object literal for the '<em><b>Saving Resource</b></em>' attribute
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Saving Resource</b></em>' attribute feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EAttribute PROJECT__SAVING_RESOURCE = eINSTANCE.getProject_SavingResource();
@@ -4272,7 +3897,6 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Revision</b></em>' reference feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EReference PROJECT__REVISION = eINSTANCE.getProject_Revision();
@@ -4286,17 +3910,15 @@ public interface ClassMakerPackage extends EPackage {
 		EReference PROJECT__REVISIONS = eINSTANCE.getProject_Revisions();
 
 		/**
-		 * The meta object literal for the '<em><b>Project Version</b></em>' attribute
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Project Version</b></em>' attribute feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EAttribute PROJECT__PROJECT_VERSION = eINSTANCE.getProject_ProjectVersion();
 
 		/**
-		 * The meta object literal for the '<em><b>Select Reveal Handler</b></em>'
-		 * reference feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Select Reveal Handler</b></em>' reference feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EReference PROJECT__SELECT_REVEAL_HANDLER = eINSTANCE.getProject_SelectRevealHandler();
@@ -4304,7 +3926,6 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Version</b></em>' attribute feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EAttribute PROJECT__VERSION = eINSTANCE.getProject_Version();
@@ -4312,24 +3933,20 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>State</b></em>' reference feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EReference PROJECT__STATE = eINSTANCE.getProject_State();
 
 		/**
-		 * The meta object literal for the '<em><b>Model Resource Adapter</b></em>'
-		 * containment reference feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Model Resource Adapter</b></em>' containment reference feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EReference PROJECT__MODEL_RESOURCE_ADAPTER = eINSTANCE.getProject_ModelResourceAdapter();
 
 		/**
-		 * The meta object literal for the
-		 * '{@link org.enterprisedomain.classmaker.impl.ModelPairImpl <em>Model
-		 * Pair</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '{@link org.enterprisedomain.classmaker.impl.ModelPairImpl <em>Model Pair</em>}' class.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @see org.enterprisedomain.classmaker.impl.ModelPairImpl
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getModelPair()
 		 * @generated
@@ -4337,34 +3954,29 @@ public interface ClassMakerPackage extends EPackage {
 		EClass MODEL_PAIR = eINSTANCE.getModelPair();
 
 		/**
-		 * The meta object literal for the '<em><b>Dynamic</b></em>' containment
-		 * reference feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Dynamic</b></em>' containment reference feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EReference MODEL_PAIR__DYNAMIC = eINSTANCE.getModelPair_Dynamic();
 
 		/**
-		 * The meta object literal for the '<em><b>Generated</b></em>' reference
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Generated</b></em>' reference feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EReference MODEL_PAIR__GENERATED = eINSTANCE.getModelPair_Generated();
 
 		/**
-		 * The meta object literal for the '<em><b>Parent</b></em>' container reference
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Parent</b></em>' container reference feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EReference MODEL_PAIR__PARENT = eINSTANCE.getModelPair_Parent();
 
 		/**
-		 * The meta object literal for the
-		 * '{@link org.enterprisedomain.classmaker.impl.SCMOperatorImpl <em>SCM
-		 * Operator</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '{@link org.enterprisedomain.classmaker.impl.SCMOperatorImpl <em>SCM Operator</em>}' class.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @see org.enterprisedomain.classmaker.impl.SCMOperatorImpl
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getSCMOperator()
 		 * @generated
@@ -4372,9 +3984,8 @@ public interface ClassMakerPackage extends EPackage {
 		EClass SCM_OPERATOR = eINSTANCE.getSCMOperator();
 
 		/**
-		 * The meta object literal for the '<em><b>Project Name</b></em>' attribute
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Project Name</b></em>' attribute feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EAttribute SCM_OPERATOR__PROJECT_NAME = eINSTANCE.getSCMOperator_ProjectName();
@@ -4382,16 +3993,13 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Registry</b></em>' reference feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EReference SCM_OPERATOR__REGISTRY = eINSTANCE.getSCMOperator_Registry();
 
 		/**
-		 * The meta object literal for the
-		 * '{@link org.enterprisedomain.classmaker.impl.SCMRegistryImpl <em>SCM
-		 * Registry</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '{@link org.enterprisedomain.classmaker.impl.SCMRegistryImpl <em>SCM Registry</em>}' class.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @see org.enterprisedomain.classmaker.impl.SCMRegistryImpl
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getSCMRegistry()
 		 * @generated
@@ -4399,10 +4007,8 @@ public interface ClassMakerPackage extends EPackage {
 		EClass SCM_REGISTRY = eINSTANCE.getSCMRegistry();
 
 		/**
-		 * The meta object literal for the
-		 * '{@link org.eclipse.emf.ecore.resource.Resource <em>Resource</em>}' class.
+		 * The meta object literal for the '{@link org.eclipse.emf.ecore.resource.Resource <em>Resource</em>}' class.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @see org.eclipse.emf.ecore.resource.Resource
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getResource()
 		 * @generated
@@ -4410,11 +4016,9 @@ public interface ClassMakerPackage extends EPackage {
 		EClass RESOURCE = eINSTANCE.getResource();
 
 		/**
-		 * The meta object literal for the
-		 * '{@link org.enterprisedomain.classmaker.impl.ResourceChangeListenerImpl
-		 * <em>Resource Change Listener</em>}' class. <!-- begin-user-doc --> <!--
+		 * The meta object literal for the '{@link org.enterprisedomain.classmaker.impl.ResourceChangeListenerImpl <em>Resource Change Listener</em>}' class.
+		 * <!-- begin-user-doc --> <!--
 		 * end-user-doc -->
-		 * 
 		 * @see org.enterprisedomain.classmaker.impl.ResourceChangeListenerImpl
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getResourceChangeListener()
 		 * @generated
@@ -4422,10 +4026,8 @@ public interface ClassMakerPackage extends EPackage {
 		EClass RESOURCE_CHANGE_LISTENER = eINSTANCE.getResourceChangeListener();
 
 		/**
-		 * The meta object literal for the
-		 * '{@link org.eclipse.emf.common.notify.Notification <em>Notification</em>}'
-		 * class. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '{@link org.eclipse.emf.common.notify.Notification <em>Notification</em>}' class.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @see org.eclipse.emf.common.notify.Notification
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getNotification()
 		 * @generated
@@ -4433,11 +4035,9 @@ public interface ClassMakerPackage extends EPackage {
 		EClass NOTIFICATION = eINSTANCE.getNotification();
 
 		/**
-		 * The meta object literal for the
-		 * '{@link org.enterprisedomain.classmaker.impl.CompletionListenerImpl
-		 * <em>Completion Listener</em>}' class. <!-- begin-user-doc --> <!--
+		 * The meta object literal for the '{@link org.enterprisedomain.classmaker.impl.CompletionListenerImpl <em>Completion Listener</em>}' class.
+		 * <!-- begin-user-doc --> <!--
 		 * end-user-doc -->
-		 * 
 		 * @see org.enterprisedomain.classmaker.impl.CompletionListenerImpl
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getCompletionListener()
 		 * @generated
@@ -4445,10 +4045,8 @@ public interface ClassMakerPackage extends EPackage {
 		EClass COMPLETION_LISTENER = eINSTANCE.getCompletionListener();
 
 		/**
-		 * The meta object literal for the
-		 * '{@link org.enterprisedomain.classmaker.jobs.Worker <em>Worker</em>}' class.
+		 * The meta object literal for the '{@link org.enterprisedomain.classmaker.jobs.Worker <em>Worker</em>}' class.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @see org.enterprisedomain.classmaker.jobs.Worker
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getWorker()
 		 * @generated
@@ -4456,11 +4054,9 @@ public interface ClassMakerPackage extends EPackage {
 		EClass WORKER = eINSTANCE.getWorker();
 
 		/**
-		 * The meta object literal for the
-		 * '{@link org.enterprisedomain.classmaker.impl.SelectRevealHandlerImpl
-		 * <em>Select Reveal Handler</em>}' class. <!-- begin-user-doc --> <!--
+		 * The meta object literal for the '{@link org.enterprisedomain.classmaker.impl.SelectRevealHandlerImpl <em>Select Reveal Handler</em>}' class.
+		 * <!-- begin-user-doc --> <!--
 		 * end-user-doc -->
-		 * 
 		 * @see org.enterprisedomain.classmaker.impl.SelectRevealHandlerImpl
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getSelectRevealHandler()
 		 * @generated
@@ -4468,10 +4064,8 @@ public interface ClassMakerPackage extends EPackage {
 		EClass SELECT_REVEAL_HANDLER = eINSTANCE.getSelectRevealHandler();
 
 		/**
-		 * The meta object literal for the
-		 * '{@link org.enterprisedomain.classmaker.impl.ResourceAdapterImpl <em>Resource
-		 * Adapter</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '{@link org.enterprisedomain.classmaker.impl.ResourceAdapterImpl <em>Resource Adapter</em>}' class.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @see org.enterprisedomain.classmaker.impl.ResourceAdapterImpl
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getResourceAdapter()
 		 * @generated
@@ -4481,7 +4075,6 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Resource</b></em>' reference feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EReference RESOURCE_ADAPTER__RESOURCE = eINSTANCE.getResourceAdapter_Resource();
@@ -4489,24 +4082,20 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Filename</b></em>' attribute feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EAttribute RESOURCE_ADAPTER__FILENAME = eINSTANCE.getResourceAdapter_Filename();
 
 		/**
-		 * The meta object literal for the '<em><b>Project</b></em>' container reference
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Project</b></em>' container reference feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EReference RESOURCE_ADAPTER__PROJECT = eINSTANCE.getResourceAdapter_Project();
 
 		/**
-		 * The meta object literal for the
-		 * '{@link org.enterprisedomain.classmaker.impl.ClassMakerServiceImpl
-		 * <em>Service</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '{@link org.enterprisedomain.classmaker.impl.ClassMakerServiceImpl <em>Service</em>}' class.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerServiceImpl
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getClassMakerService()
 		 * @generated
@@ -4514,18 +4103,15 @@ public interface ClassMakerPackage extends EPackage {
 		EClass CLASS_MAKER_SERVICE = eINSTANCE.getClassMakerService();
 
 		/**
-		 * The meta object literal for the '<em><b>Workspace</b></em>' containment
-		 * reference feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Workspace</b></em>' containment reference feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EReference CLASS_MAKER_SERVICE__WORKSPACE = eINSTANCE.getClassMakerService_Workspace();
 
 		/**
-		 * The meta object literal for the
-		 * '{@link org.enterprisedomain.classmaker.impl.BlueprintImpl
-		 * <em>Blueprint</em>}' class. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '{@link org.enterprisedomain.classmaker.impl.BlueprintImpl <em>Blueprint</em>}' class.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @see org.enterprisedomain.classmaker.impl.BlueprintImpl
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getBlueprint()
 		 * @generated
@@ -4533,25 +4119,22 @@ public interface ClassMakerPackage extends EPackage {
 		EClass BLUEPRINT = eINSTANCE.getBlueprint();
 
 		/**
-		 * The meta object literal for the '<em><b>Dynamic Model</b></em>' reference
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Dynamic Model</b></em>' reference feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EReference BLUEPRINT__DYNAMIC_MODEL = eINSTANCE.getBlueprint_DynamicModel();
 
 		/**
-		 * The meta object literal for the '<em><b>Dependencies</b></em>' attribute list
-		 * feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Dependencies</b></em>' attribute list feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EAttribute BLUEPRINT__DEPENDENCIES = eINSTANCE.getBlueprint_Dependencies();
 
 		/**
-		 * The meta object literal for the '<em><b>Completion Listeners</b></em>'
-		 * reference list feature. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em><b>Completion Listeners</b></em>' reference list feature.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @generated
 		 */
 		EReference BLUEPRINT__COMPLETION_LISTENERS = eINSTANCE.getBlueprint_CompletionListeners();
@@ -4559,7 +4142,6 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Edit</b></em>' attribute feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EAttribute BLUEPRINT__EDIT = eINSTANCE.getBlueprint_Edit();
@@ -4567,17 +4149,14 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Editor</b></em>' attribute feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EAttribute BLUEPRINT__EDITOR = eINSTANCE.getBlueprint_Editor();
 
 		/**
-		 * The meta object literal for the
-		 * '{@link org.enterprisedomain.classmaker.impl.CompletionNotificationAdapterImpl
-		 * <em>Completion Notification Adapter</em>}' class. <!-- begin-user-doc -->
+		 * The meta object literal for the '{@link org.enterprisedomain.classmaker.impl.CompletionNotificationAdapterImpl <em>Completion Notification Adapter</em>}' class.
+		 * <!-- begin-user-doc -->
 		 * <!-- end-user-doc -->
-		 * 
 		 * @see org.enterprisedomain.classmaker.impl.CompletionNotificationAdapterImpl
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getCompletionNotificationAdapter()
 		 * @generated
@@ -4587,7 +4166,6 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Error</b></em>' attribute feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EAttribute COMPLETION_NOTIFICATION_ADAPTER__ERROR = eINSTANCE.getCompletionNotificationAdapter_Error();
@@ -4595,7 +4173,6 @@ public interface ClassMakerPackage extends EPackage {
 		/**
 		 * The meta object literal for the '<em><b>Project</b></em>' reference feature.
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		EReference COMPLETION_NOTIFICATION_ADAPTER__PROJECT = eINSTANCE.getCompletionNotificationAdapter_Project();
@@ -4611,9 +4188,8 @@ public interface ClassMakerPackage extends EPackage {
 		EDataType LOCALE = eINSTANCE.getLocale();
 
 		/**
-		 * The meta object literal for the '{@link org.enterprisedomain.classmaker.Stage
-		 * <em>Stage</em>}' enum. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '{@link org.enterprisedomain.classmaker.Stage <em>Stage</em>}' enum.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @see org.enterprisedomain.classmaker.Stage
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getStage()
 		 * @generated
@@ -4671,9 +4247,8 @@ public interface ClassMakerPackage extends EPackage {
 		EDataType CORE_EXCEPTION = eINSTANCE.getCoreException();
 
 		/**
-		 * The meta object literal for the '<em>Invocation Target Exception</em>' data
-		 * type. <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
+		 * The meta object literal for the '<em>Invocation Target Exception</em>' data type.
+		 * <!-- begin-user-doc --> <!-- end-user-doc -->
 		 * @see java.lang.reflect.InvocationTargetException
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getInvocationTargetException()
 		 * @generated
@@ -4691,9 +4266,9 @@ public interface ClassMakerPackage extends EPackage {
 		EDataType ISTATUS = eINSTANCE.getIStatus();
 
 		/**
-		 * The meta object literal for the '<em>URI</em>' data type. <!-- begin-user-doc
+		 * The meta object literal for the '<em>URI</em>' data type.
+		 * <!-- begin-user-doc
 		 * --> <!-- end-user-doc -->
-		 * 
 		 * @see org.eclipse.emf.common.util.URI
 		 * @see org.enterprisedomain.classmaker.impl.ClassMakerPackageImpl#getURI()
 		 * @generated
