+++ bundles/org.enterprisedomain.classmaker/src/org/enterprisedomain/classmaker/ClassMakerService.java
@@ -38,15 +38,15 @@ import org.osgi.framework.Version;
  * <!-- begin-user-doc --> A representation of the model object
  * '<em><b>Service</b></em>'. <!-- end-user-doc -->
  *
- * <!-- begin-model-doc --> ClassMaker API facade service. <!-- end-model-doc
- * -->
+ * <!-- begin-model-doc -->
+ * ClassMaker API facade service.
+ * <!-- end-model-doc -->
  *
  * <p>
  * The following features are supported:
  * </p>
  * <ul>
- * <li>{@link org.enterprisedomain.classmaker.ClassMakerService#getWorkspace
- * <em>Workspace</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.ClassMakerService#getWorkspace <em>Workspace</em>}</li>
  * </ul>
  *
  * @see org.enterprisedomain.classmaker.ClassMakerPackage#getClassMakerService()
@@ -143,11 +143,9 @@ public interface ClassMakerService extends EObject {
 	Workspace getWorkspace();
 
 	/**
-	 * Sets the value of the
-	 * '{@link org.enterprisedomain.classmaker.ClassMakerService#getWorkspace
-	 * <em>Workspace</em>}' containment reference. <!-- begin-user-doc --> <!--
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.ClassMakerService#getWorkspace <em>Workspace</em>}' containment reference.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @param value the new value of the '<em>Workspace</em>' containment reference.
 	 * @see #getWorkspace()
 	 * @generated
@@ -182,27 +180,21 @@ public interface ClassMakerService extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model exceptions="org.enterprisedomain.classmaker.CoreException"
-	 *        monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
+	 * @model exceptions="org.enterprisedomain.classmaker.CoreException" monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
 	 * @generated
 	 */
 	EObject replace(EObject source, EObject target, IProgressMonitor monitor) throws CoreException;
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model exceptions="org.enterprisedomain.classmaker.CoreException"
-	 *        monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
+	 * @model exceptions="org.enterprisedomain.classmaker.CoreException" monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
 	 * @generated
 	 */
 	EObject replace(Blueprint source, Blueprint target, IProgressMonitor monitor) throws CoreException;
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model exceptions="org.enterprisedomain.classmaker.CoreException"
-	 *        monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
+	 * @model exceptions="org.enterprisedomain.classmaker.CoreException" monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
 	 * @generated
 	 */
 	EObject replace(EObject source, EObject target, boolean changeVersion, IProgressMonitor monitor)
@@ -210,9 +202,7 @@ public interface ClassMakerService extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model exceptions="org.enterprisedomain.classmaker.CoreException"
-	 *        monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
+	 * @model exceptions="org.enterprisedomain.classmaker.CoreException" monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
 	 * @generated
 	 */
 	EObject replace(Blueprint source, Blueprint target, boolean changeVersion, IProgressMonitor monitor)
@@ -220,20 +210,14 @@ public interface ClassMakerService extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model exceptions="org.enterprisedomain.classmaker.CoreException"
-	 *        versionDataType="org.enterprisedomain.classmaker.OSGiVersion"
-	 *        monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
+	 * @model exceptions="org.enterprisedomain.classmaker.CoreException" versionDataType="org.enterprisedomain.classmaker.OSGiVersion" monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
 	 * @generated
 	 */
 	EObject replace(EObject source, EObject target, Version version, IProgressMonitor monitor) throws CoreException;
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model exceptions="org.enterprisedomain.classmaker.CoreException"
-	 *        versionDataType="org.enterprisedomain.classmaker.OSGiVersion"
-	 *        monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
+	 * @model exceptions="org.enterprisedomain.classmaker.CoreException" versionDataType="org.enterprisedomain.classmaker.OSGiVersion" monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
 	 * @generated
 	 */
 	EObject replace(Blueprint source, Blueprint target, Version version, IProgressMonitor monitor) throws CoreException;
@@ -272,12 +256,7 @@ public interface ClassMakerService extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model type="org.enterprisedomain.classmaker.Future&lt;? extends
-	 *        org.eclipse.emf.ecore.EObject&gt;"
-	 *        exceptions="org.enterprisedomain.classmaker.CoreException"
-	 *        executorType="org.enterprisedomain.classmaker.Executor"
-	 *        monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
+	 * @model type="org.enterprisedomain.classmaker.Future&lt;? extends org.eclipse.emf.ecore.EObject&gt;" exceptions="org.enterprisedomain.classmaker.CoreException" executorType="org.enterprisedomain.classmaker.Executor" monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
 	 * @generated
 	 */
 	Future<? extends EObject> replace(EObject source, EObject target, Executor executor, IProgressMonitor monitor)
@@ -285,12 +264,7 @@ public interface ClassMakerService extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model type="org.enterprisedomain.classmaker.Future&lt;? extends
-	 *        org.eclipse.emf.ecore.EObject&gt;"
-	 *        exceptions="org.enterprisedomain.classmaker.CoreException"
-	 *        executorType="org.enterprisedomain.classmaker.Executor"
-	 *        monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
+	 * @model type="org.enterprisedomain.classmaker.Future&lt;? extends org.eclipse.emf.ecore.EObject&gt;" exceptions="org.enterprisedomain.classmaker.CoreException" executorType="org.enterprisedomain.classmaker.Executor" monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
 	 * @generated
 	 */
 	Future<? extends EObject> replace(Blueprint source, Blueprint target, Executor executor, IProgressMonitor monitor)
@@ -298,12 +272,7 @@ public interface ClassMakerService extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model type="org.enterprisedomain.classmaker.Future&lt;? extends
-	 *        org.eclipse.emf.ecore.EObject&gt;"
-	 *        exceptions="org.enterprisedomain.classmaker.CoreException"
-	 *        executorType="org.enterprisedomain.classmaker.Executor"
-	 *        monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
+	 * @model type="org.enterprisedomain.classmaker.Future&lt;? extends org.eclipse.emf.ecore.EObject&gt;" exceptions="org.enterprisedomain.classmaker.CoreException" executorType="org.enterprisedomain.classmaker.Executor" monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
 	 * @generated
 	 */
 	Future<? extends EObject> replace(EObject source, EObject target, boolean changeVersion, Executor executor,
@@ -311,12 +280,7 @@ public interface ClassMakerService extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model type="org.enterprisedomain.classmaker.Future&lt;? extends
-	 *        org.eclipse.emf.ecore.EObject&gt;"
-	 *        exceptions="org.enterprisedomain.classmaker.CoreException"
-	 *        executorType="org.enterprisedomain.classmaker.Executor"
-	 *        monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
+	 * @model type="org.enterprisedomain.classmaker.Future&lt;? extends org.eclipse.emf.ecore.EObject&gt;" exceptions="org.enterprisedomain.classmaker.CoreException" executorType="org.enterprisedomain.classmaker.Executor" monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
 	 * @generated
 	 */
 	Future<? extends EObject> replace(Blueprint source, Blueprint target, boolean changeVersion, Executor executor,
@@ -324,13 +288,7 @@ public interface ClassMakerService extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model type="org.enterprisedomain.classmaker.Future&lt;? extends
-	 *        org.eclipse.emf.ecore.EObject&gt;"
-	 *        exceptions="org.enterprisedomain.classmaker.CoreException"
-	 *        versionDataType="org.enterprisedomain.classmaker.OSGiVersion"
-	 *        executorType="org.enterprisedomain.classmaker.Executor"
-	 *        monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
+	 * @model type="org.enterprisedomain.classmaker.Future&lt;? extends org.eclipse.emf.ecore.EObject&gt;" exceptions="org.enterprisedomain.classmaker.CoreException" versionDataType="org.enterprisedomain.classmaker.OSGiVersion" executorType="org.enterprisedomain.classmaker.Executor" monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
 	 * @generated
 	 */
 	Future<? extends EObject> replace(EObject source, EObject target, Version version, Executor executor,
@@ -338,13 +296,7 @@ public interface ClassMakerService extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model type="org.enterprisedomain.classmaker.Future&lt;? extends
-	 *        org.eclipse.emf.ecore.EObject&gt;"
-	 *        exceptions="org.enterprisedomain.classmaker.CoreException"
-	 *        versionDataType="org.enterprisedomain.classmaker.OSGiVersion"
-	 *        executorType="org.enterprisedomain.classmaker.Executor"
-	 *        monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
+	 * @model type="org.enterprisedomain.classmaker.Future&lt;? extends org.eclipse.emf.ecore.EObject&gt;" exceptions="org.enterprisedomain.classmaker.CoreException" versionDataType="org.enterprisedomain.classmaker.OSGiVersion" executorType="org.enterprisedomain.classmaker.Executor" monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
 	 * @generated
 	 */
 	Future<? extends EObject> replace(Blueprint source, Blueprint target, Version version, Executor executor,
@@ -352,44 +304,34 @@ public interface ClassMakerService extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model exceptions="org.enterprisedomain.classmaker.CoreException"
-	 *        transformationURIDataType="org.enterprisedomain.classmaker.URI"
+	 * @model exceptions="org.enterprisedomain.classmaker.CoreException" transformationURIDataType="org.enterprisedomain.classmaker.URI"
 	 * @generated
 	 */
 	EObject transform(EObject source, URI transformationURI) throws CoreException;
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model exceptions="org.enterprisedomain.classmaker.CoreException"
-	 *        transformationURIDataType="org.enterprisedomain.classmaker.URI"
-	 *        monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
+	 * @model exceptions="org.enterprisedomain.classmaker.CoreException" transformationURIDataType="org.enterprisedomain.classmaker.URI" monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
 	 * @generated
 	 */
 	EObject transform(EObject source, URI transformationURI, IProgressMonitor monitor) throws CoreException;
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model exceptions="org.enterprisedomain.classmaker.InvocationTargetException"
-	 *        argumentsMany="true"
+	 * @model exceptions="org.enterprisedomain.classmaker.InvocationTargetException" argumentsMany="true"
 	 * @generated
 	 */
 	Object invoke(EOperation operation, EObject object, EList<Object> arguments) throws InvocationTargetException;
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model exceptions="org.enterprisedomain.classmaker.CoreException"
-	 *        monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
+	 * @model exceptions="org.enterprisedomain.classmaker.CoreException" monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
 	 * @generated
 	 */
 	void delete(String packageName, IProgressMonitor monitor) throws CoreException;
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -397,7 +339,6 @@ public interface ClassMakerService extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -405,7 +346,6 @@ public interface ClassMakerService extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -413,7 +353,6 @@ public interface ClassMakerService extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -421,7 +360,6 @@ public interface ClassMakerService extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -429,7 +367,6 @@ public interface ClassMakerService extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
