+++ bundles/org.enterprisedomain.classmaker/src/org/enterprisedomain/classmaker/CompletionListener.java
@@ -21,8 +21,9 @@ import org.eclipse.emf.ecore.EObject;
  * <!-- begin-user-doc --> A representation of the model object
  * '<em><b>Completion Listener</b></em>'. <!-- end-user-doc -->
  *
- * <!-- begin-model-doc --> A process completion listener. <!-- end-model-doc
- * -->
+ * <!-- begin-model-doc -->
+ * A process completion listener.
+ * <!-- end-model-doc -->
  *
  *
  * @see org.enterprisedomain.classmaker.ClassMakerPackage#getCompletionListener()
