+++ bundles/org.enterprisedomain.classmaker/src/org/enterprisedomain/classmaker/Project.java
@@ -27,51 +27,35 @@ import org.osgi.framework.Version;
  * <!-- begin-user-doc --> A representation of the model object '
  * <em><b>Project</b></em>'. <!-- end-user-doc -->
  *
- * <!-- begin-model-doc --> Default project class. <!-- end-model-doc -->
+ * <!-- begin-model-doc -->
+ * Default project class.
+ * <!-- end-model-doc -->
  *
  * <p>
  * The following features are supported:
  * </p>
  * <ul>
- * <li>{@link org.enterprisedomain.classmaker.Project#getName
- * <em>Name</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.Project#getProjectName <em>Project
- * Name</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.Project#getChildren
- * <em>Children</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.Project#isDirty
- * <em>Dirty</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.Project#getWorkspace
- * <em>Workspace</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.Project#getResourcePath
- * <em>Resource Path</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.Project#isNeedCompletionNotification
- * <em>Need Completion Notification</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.Project#getCompletionNotificationAdapter
- * <em>Completion Notification Adapter</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.Project#getResourceReloadListener
- * <em>Resource Reload Listener</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.Project#isSavingResource
- * <em>Saving Resource</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.Project#getRevision
- * <em>Revision</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.Project#getRevisions
- * <em>Revisions</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.Project#getProjectVersion
- * <em>Project Version</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.Project#getSelectRevealHandler
- * <em>Select Reveal Handler</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.Project#getVersion
- * <em>Version</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.Project#getState
- * <em>State</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.Project#getModelResourceAdapter
- * <em>Model Resource Adapter</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.Project#getName <em>Name</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.Project#getProjectName <em>Project Name</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.Project#getChildren <em>Children</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.Project#isDirty <em>Dirty</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.Project#getWorkspace <em>Workspace</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.Project#getResourcePath <em>Resource Path</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.Project#isNeedCompletionNotification <em>Need Completion Notification</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.Project#getCompletionNotificationAdapter <em>Completion Notification Adapter</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.Project#getResourceReloadListener <em>Resource Reload Listener</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.Project#isSavingResource <em>Saving Resource</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.Project#getRevision <em>Revision</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.Project#getRevisions <em>Revisions</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.Project#getProjectVersion <em>Project Version</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.Project#getSelectRevealHandler <em>Select Reveal Handler</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.Project#getVersion <em>Version</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.Project#getState <em>State</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.Project#getModelResourceAdapter <em>Model Resource Adapter</em>}</li>
  * </ul>
  *
  * @see org.enterprisedomain.classmaker.ClassMakerPackage#getProject()
- * @model superTypes="org.enterprisedomain.classmaker.ISchedulingRule
- *        org.enterprisedomain.classmaker.Item"
+ * @model superTypes="org.enterprisedomain.classmaker.ISchedulingRule org.enterprisedomain.classmaker.Item"
  * @generated
  */
 public interface Project extends ISchedulingRule, Item {
@@ -94,9 +78,8 @@ public interface Project extends ISchedulingRule, Item {
 	String getName();
 
 	/**
-	 * Sets the value of the '{@link org.enterprisedomain.classmaker.Project#getName
-	 * <em>Name</em>}' attribute. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.Project#getName <em>Name</em>}' attribute.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @param value the new value of the '<em>Name</em>' attribute.
 	 * @see #getName()
 	 * @generated
@@ -123,10 +106,8 @@ public interface Project extends ISchedulingRule, Item {
 	String getProjectName();
 
 	/**
-	 * Sets the value of the
-	 * '{@link org.enterprisedomain.classmaker.Project#getProjectName <em>Project
-	 * Name</em>}' attribute. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.Project#getProjectName <em>Project Name</em>}' attribute.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @param value the new value of the '<em>Project Name</em>' attribute.
 	 * @see #getProjectName()
 	 * @generated
@@ -167,9 +148,8 @@ public interface Project extends ISchedulingRule, Item {
 	boolean isDirty();
 
 	/**
-	 * Sets the value of the '{@link org.enterprisedomain.classmaker.Project#isDirty
-	 * <em>Dirty</em>}' attribute. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.Project#isDirty <em>Dirty</em>}' attribute.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @param value the new value of the '<em>Dirty</em>' attribute.
 	 * @see #isDirty()
 	 * @generated
@@ -199,11 +179,9 @@ public interface Project extends ISchedulingRule, Item {
 	Workspace getWorkspace();
 
 	/**
-	 * Sets the value of the
-	 * '{@link org.enterprisedomain.classmaker.Project#getWorkspace
-	 * <em>Workspace</em>}' container reference. <!-- begin-user-doc --> <!--
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.Project#getWorkspace <em>Workspace</em>}' container reference.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @param value the new value of the '<em>Workspace</em>' container reference.
 	 * @see #getWorkspace()
 	 * @generated
@@ -241,13 +219,10 @@ public interface Project extends ISchedulingRule, Item {
 	boolean isNeedCompletionNotification();
 
 	/**
-	 * Sets the value of the
-	 * '{@link org.enterprisedomain.classmaker.Project#isNeedCompletionNotification
-	 * <em>Need Completion Notification</em>}' attribute. <!-- begin-user-doc -->
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.Project#isNeedCompletionNotification <em>Need Completion Notification</em>}' attribute.
+	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
-	 * 
-	 * @param value the new value of the '<em>Need Completion Notification</em>'
-	 *              attribute.
+	 * @param value the new value of the '<em>Need Completion Notification</em>' attribute.
 	 * @see #isNeedCompletionNotification()
 	 * @generated
 	 */
@@ -287,14 +262,13 @@ public interface Project extends ISchedulingRule, Item {
 	void setCompletionNotificationAdapter(CompletionNotificationAdapter value);
 
 	/**
-	 * Returns the value of the '<em><b>Resource Reload Listener</b></em>'
-	 * reference. <!-- begin-user-doc -->
+	 * Returns the value of the '<em><b>Resource Reload Listener</b></em>' reference.
+	 * <!-- begin-user-doc -->
 	 * <p>
 	 * If the meaning of the '<em>Resource Reload Listener</em>' reference isn't
 	 * clear, there really should be more of a description here...
 	 * </p>
 	 * <!-- end-user-doc -->
-	 * 
 	 * @return the value of the '<em>Resource Reload Listener</em>' reference.
 	 * @see org.enterprisedomain.classmaker.ClassMakerPackage#getProject_ResourceReloadListener()
 	 * @model resolveProxies="false" changeable="false"
@@ -303,14 +277,14 @@ public interface Project extends ISchedulingRule, Item {
 	ResourceChangeListener getResourceReloadListener();
 
 	/**
-	 * Returns the value of the '<em><b>Saving Resource</b></em>' attribute. The
-	 * default value is <code>"false"</code>. <!-- begin-user-doc -->
+	 * Returns the value of the '<em><b>Saving Resource</b></em>' attribute.
+	 * The default value is <code>"false"</code>.
+	 * <!-- begin-user-doc -->
 	 * <p>
 	 * If the meaning of the '<em>Saving Resource</em>' attribute isn't clear, there
 	 * really should be more of a description here...
 	 * </p>
 	 * <!-- end-user-doc -->
-	 * 
 	 * @return the value of the '<em>Saving Resource</em>' attribute.
 	 * @see #setSavingResource(boolean)
 	 * @see org.enterprisedomain.classmaker.ClassMakerPackage#getProject_SavingResource()
@@ -320,10 +294,8 @@ public interface Project extends ISchedulingRule, Item {
 	boolean isSavingResource();
 
 	/**
-	 * Sets the value of the
-	 * '{@link org.enterprisedomain.classmaker.Project#isSavingResource <em>Saving
-	 * Resource</em>}' attribute. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.Project#isSavingResource <em>Saving Resource</em>}' attribute.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @param value the new value of the '<em>Saving Resource</em>' attribute.
 	 * @see #isSavingResource()
 	 * @generated
@@ -344,10 +316,8 @@ public interface Project extends ISchedulingRule, Item {
 	Revision getRevision();
 
 	/**
-	 * Sets the value of the
-	 * '{@link org.enterprisedomain.classmaker.Project#getRevision
-	 * <em>Revision</em>}' reference. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.Project#getRevision <em>Revision</em>}' reference.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @param value the new value of the '<em>Revision</em>' reference.
 	 * @see #getRevision()
 	 * @generated
@@ -385,10 +355,8 @@ public interface Project extends ISchedulingRule, Item {
 	Version getProjectVersion();
 
 	/**
-	 * Sets the value of the
-	 * '{@link org.enterprisedomain.classmaker.Project#getProjectVersion <em>Project
-	 * Version</em>}' attribute. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.Project#getProjectVersion <em>Project Version</em>}' attribute.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @param value the new value of the '<em>Project Version</em>' attribute.
 	 * @see #getProjectVersion()
 	 * @generated
@@ -403,7 +371,6 @@ public interface Project extends ISchedulingRule, Item {
 	 * there really should be more of a description here...
 	 * </p>
 	 * <!-- end-user-doc -->
-	 * 
 	 * @return the value of the '<em>Select Reveal Handler</em>' reference.
 	 * @see #setSelectRevealHandler(SelectRevealHandler)
 	 * @see org.enterprisedomain.classmaker.ClassMakerPackage#getProject_SelectRevealHandler()
@@ -413,11 +380,9 @@ public interface Project extends ISchedulingRule, Item {
 	SelectRevealHandler getSelectRevealHandler();
 
 	/**
-	 * Sets the value of the
-	 * '{@link org.enterprisedomain.classmaker.Project#getSelectRevealHandler
-	 * <em>Select Reveal Handler</em>}' reference. <!-- begin-user-doc --> <!--
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.Project#getSelectRevealHandler <em>Select Reveal Handler</em>}' reference.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @param value the new value of the '<em>Select Reveal Handler</em>' reference.
 	 * @see #getSelectRevealHandler()
 	 * @generated
@@ -439,10 +404,8 @@ public interface Project extends ISchedulingRule, Item {
 	Version getVersion();
 
 	/**
-	 * Sets the value of the
-	 * '{@link org.enterprisedomain.classmaker.Project#getVersion <em>Version</em>}'
-	 * attribute. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.Project#getVersion <em>Version</em>}' attribute.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @param value the new value of the '<em>Version</em>' attribute.
 	 * @see #getVersion()
 	 * @generated
@@ -480,25 +443,28 @@ public interface Project extends ISchedulingRule, Item {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model exceptions="org.enterprisedomain.classmaker.CoreException"
-	 *        monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
+	 * @model exceptions="org.enterprisedomain.classmaker.CoreException" monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
 	 * @generated
 	 */
 	void create(IProgressMonitor monitor) throws CoreException;
 
 	/**
+	 * <!-- begin-user-doc -->
+	 * <!-- end-user-doc -->
+	 * @model exceptions="org.enterprisedomain.classmaker.CoreException" monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
+	 * @generated
+	 */
+	Revision createRevision(IProgressMonitor monitor) throws CoreException;
+
+	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model exceptions="org.enterprisedomain.classmaker.CoreException"
-	 *        monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
+	 * @model exceptions="org.enterprisedomain.classmaker.CoreException" monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
 	 * @generated
 	 */
 	void delete(IProgressMonitor monitor) throws CoreException;
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model objectsMany="true"
 	 * @generated
 	 */
@@ -506,27 +472,21 @@ public interface Project extends ISchedulingRule, Item {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model exceptions="org.enterprisedomain.classmaker.CoreException"
-	 *        monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
+	 * @model exceptions="org.enterprisedomain.classmaker.CoreException" monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
 	 * @generated
 	 */
 	String make(IProgressMonitor monitor) throws CoreException;
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model exceptions="org.enterprisedomain.classmaker.CoreException"
-	 *        monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
+	 * @model exceptions="org.enterprisedomain.classmaker.CoreException" monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
 	 * @generated
 	 */
 	boolean open(IProgressMonitor monitor) throws CoreException;
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model exceptions="org.enterprisedomain.classmaker.CoreException"
-	 *        monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
+	 * @model exceptions="org.enterprisedomain.classmaker.CoreException" monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
 	 * @generated
 	 */
 	void close(IProgressMonitor monitor) throws CoreException;
@@ -545,7 +505,6 @@ public interface Project extends ISchedulingRule, Item {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -553,7 +512,6 @@ public interface Project extends ISchedulingRule, Item {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -561,7 +519,6 @@ public interface Project extends ISchedulingRule, Item {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model exceptions="org.enterprisedomain.classmaker.Exception"
 	 * @generated
 	 */
@@ -569,16 +526,13 @@ public interface Project extends ISchedulingRule, Item {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model exceptions="org.enterprisedomain.classmaker.Exception"
-	 *        notificationType="org.enterprisedomain.classmaker.Notification"
+	 * @model exceptions="org.enterprisedomain.classmaker.Exception" notificationType="org.enterprisedomain.classmaker.Notification"
 	 * @generated
 	 */
 	void notifyResourceChanged(Notification notification) throws Exception;
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -586,7 +540,6 @@ public interface Project extends ISchedulingRule, Item {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -594,7 +547,6 @@ public interface Project extends ISchedulingRule, Item {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model versionDataType="org.enterprisedomain.classmaker.OSGiVersion"
 	 * @generated
 	 */
@@ -602,7 +554,6 @@ public interface Project extends ISchedulingRule, Item {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model versionDataType="org.enterprisedomain.classmaker.OSGiVersion"
 	 * @generated
 	 */
@@ -610,7 +561,6 @@ public interface Project extends ISchedulingRule, Item {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -618,10 +568,7 @@ public interface Project extends ISchedulingRule, Item {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model dataType="org.enterprisedomain.classmaker.OSGiVersion"
-	 *        exceptions="org.enterprisedomain.classmaker.CoreException"
-	 *        baseDataType="org.enterprisedomain.classmaker.OSGiVersion"
+	 * @model dataType="org.enterprisedomain.classmaker.OSGiVersion" exceptions="org.enterprisedomain.classmaker.CoreException" baseDataType="org.enterprisedomain.classmaker.OSGiVersion"
 	 * @generated
 	 */
 	Version newVersion(Version base, boolean incrementMajor, boolean incrementMinor, boolean incrementMicro)
@@ -629,25 +576,20 @@ public interface Project extends ISchedulingRule, Item {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model dataType="org.enterprisedomain.classmaker.OSGiVersion"
-	 *        exceptions="org.enterprisedomain.classmaker.CoreException"
+	 * @model dataType="org.enterprisedomain.classmaker.OSGiVersion" exceptions="org.enterprisedomain.classmaker.CoreException"
 	 * @generated
 	 */
 	Version newVersion(boolean incrementMajor, boolean incrementMinor, boolean incrementMicro) throws CoreException;
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model dataType="org.enterprisedomain.classmaker.OSGiVersion"
-	 *        exceptions="org.enterprisedomain.classmaker.CoreException"
+	 * @model dataType="org.enterprisedomain.classmaker.OSGiVersion" exceptions="org.enterprisedomain.classmaker.CoreException"
 	 * @generated
 	 */
 	Version nextVersion() throws CoreException;
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model versionDataType="org.enterprisedomain.classmaker.OSGiVersion"
 	 * @generated
 	 */
@@ -655,7 +597,6 @@ public interface Project extends ISchedulingRule, Item {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model versionDataType="org.enterprisedomain.classmaker.OSGiVersion"
 	 * @generated
 	 */
@@ -663,7 +604,6 @@ public interface Project extends ISchedulingRule, Item {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model versionDataType="org.enterprisedomain.classmaker.OSGiVersion"
 	 * @generated
 	 */
@@ -671,7 +611,6 @@ public interface Project extends ISchedulingRule, Item {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -679,7 +618,6 @@ public interface Project extends ISchedulingRule, Item {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -687,7 +625,6 @@ public interface Project extends ISchedulingRule, Item {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -695,7 +632,6 @@ public interface Project extends ISchedulingRule, Item {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
