+++ bundles/org.enterprisedomain.classmaker/src/org/enterprisedomain/classmaker/ResourceAdapter.java
@@ -23,21 +23,17 @@ import org.eclipse.emf.ecore.resource.Resource;
  * <!-- begin-user-doc --> A representation of the model object '<em><b>Resource
  * Adapter</b></em>'. <!-- end-user-doc -->
  *
- * <!-- begin-model-doc --> {@link org.eclipse.emf.ecore.EObject EObject}
- * {@link org.eclipse.emf.common.notify.Adapter adapter} for
- * {@link org.eclipse.emf.ecore.resource.Resource Resource}. <!-- end-model-doc
- * -->
+ * <!-- begin-model-doc -->
+ * {@link org.eclipse.emf.ecore.EObject EObject} {@link org.eclipse.emf.common.notify.Adapter adapter} for {@link org.eclipse.emf.ecore.resource.Resource Resource}.
+ * <!-- end-model-doc -->
  *
  * <p>
  * The following features are supported:
  * </p>
  * <ul>
- * <li>{@link org.enterprisedomain.classmaker.ResourceAdapter#getResource
- * <em>Resource</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.ResourceAdapter#getFilename
- * <em>Filename</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.ResourceAdapter#getProject
- * <em>Project</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.ResourceAdapter#getResource <em>Resource</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.ResourceAdapter#getFilename <em>Filename</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.ResourceAdapter#getProject <em>Project</em>}</li>
  * </ul>
  *
  * @see org.enterprisedomain.classmaker.ClassMakerPackage#getResourceAdapter()
@@ -64,10 +60,8 @@ public interface ResourceAdapter extends EObject, Adapter {
 	Resource getResource();
 
 	/**
-	 * Sets the value of the
-	 * '{@link org.enterprisedomain.classmaker.ResourceAdapter#getResource
-	 * <em>Resource</em>}' reference. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.ResourceAdapter#getResource <em>Resource</em>}' reference.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @param value the new value of the '<em>Resource</em>' reference.
 	 * @see #getResource()
 	 * @generated
@@ -91,16 +85,14 @@ public interface ResourceAdapter extends EObject, Adapter {
 	String getFilename();
 
 	/**
-	 * Returns the value of the '<em><b>Project</b></em>' container reference. It is
-	 * bidirectional and its opposite is
-	 * '{@link org.enterprisedomain.classmaker.Project#getModelResourceAdapter
-	 * <em>Model Resource Adapter</em>}'. <!-- begin-user-doc -->
+	 * Returns the value of the '<em><b>Project</b></em>' container reference.
+	 * It is bidirectional and its opposite is '{@link org.enterprisedomain.classmaker.Project#getModelResourceAdapter <em>Model Resource Adapter</em>}'.
+	 * <!-- begin-user-doc -->
 	 * <p>
 	 * If the meaning of the '<em>Project</em>' container reference isn't clear,
 	 * there really should be more of a description here...
 	 * </p>
 	 * <!-- end-user-doc -->
-	 * 
 	 * @return the value of the '<em>Project</em>' container reference.
 	 * @see #setProject(Project)
 	 * @see org.enterprisedomain.classmaker.ClassMakerPackage#getResourceAdapter_Project()
@@ -111,11 +103,9 @@ public interface ResourceAdapter extends EObject, Adapter {
 	Project getProject();
 
 	/**
-	 * Sets the value of the
-	 * '{@link org.enterprisedomain.classmaker.ResourceAdapter#getProject
-	 * <em>Project</em>}' container reference. <!-- begin-user-doc --> <!--
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.ResourceAdapter#getProject <em>Project</em>}' container reference.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @param value the new value of the '<em>Project</em>' container reference.
 	 * @see #getProject()
 	 * @generated
