+++ bundles/org.enterprisedomain.classmaker/src/org/enterprisedomain/classmaker/ResourceChangeListener.java
@@ -30,9 +30,7 @@ import org.eclipse.emf.ecore.EObject;
 public interface ResourceChangeListener extends EObject {
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model exceptions="org.enterprisedomain.classmaker.Exception"
-	 *        notificationType="org.enterprisedomain.classmaker.Notification"
+	 * @model exceptions="org.enterprisedomain.classmaker.Exception" notificationType="org.enterprisedomain.classmaker.Notification"
 	 * @generated
 	 */
 	void changed(Notification notification) throws Exception;
