+++ bundles/org.enterprisedomain.classmaker/src/org/enterprisedomain/classmaker/SCMOperator.java
@@ -22,17 +22,16 @@ import org.osgi.framework.Version;
  * <!-- begin-user-doc --> A representation of the model object '<em><b>SCM
  * Operator</b></em>'. <!-- end-user-doc -->
  *
- * <!-- begin-model-doc --> Operator of Source Code Management system. <!--
- * end-model-doc -->
+ * <!-- begin-model-doc -->
+ * Operator of Source Code Management system.
+ * <!-- end-model-doc -->
  *
  * <p>
  * The following features are supported:
  * </p>
  * <ul>
- * <li>{@link org.enterprisedomain.classmaker.SCMOperator#getProjectName
- * <em>Project Name</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.SCMOperator#getRegistry
- * <em>Registry</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.SCMOperator#getProjectName <em>Project Name</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.SCMOperator#getRegistry <em>Registry</em>}</li>
  * </ul>
  *
  * @see org.enterprisedomain.classmaker.ClassMakerPackage#getSCMOperator()
@@ -92,10 +91,8 @@ public interface SCMOperator<T> extends EObject {
 	SCMRegistry<T> getRegistry();
 
 	/**
-	 * Sets the value of the
-	 * '{@link org.enterprisedomain.classmaker.SCMOperator#getRegistry
-	 * <em>Registry</em>}' reference. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.SCMOperator#getRegistry <em>Registry</em>}' reference.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @param value the new value of the '<em>Registry</em>' reference.
 	 * @see #getRegistry()
 	 * @generated
@@ -104,16 +101,13 @@ public interface SCMOperator<T> extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model kind="operation"
-	 *        exceptions="org.enterprisedomain.classmaker.Exception"
+	 * @model kind="operation" exceptions="org.enterprisedomain.classmaker.Exception"
 	 * @generated
 	 */
 	T getRepositorySCM() throws Exception;
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model exceptions="org.enterprisedomain.classmaker.Exception"
 	 * @generated
 	 */
@@ -121,7 +115,6 @@ public interface SCMOperator<T> extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model exceptions="org.enterprisedomain.classmaker.Exception"
 	 * @generated
 	 */
@@ -129,7 +122,6 @@ public interface SCMOperator<T> extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model exceptions="org.enterprisedomain.classmaker.Exception"
 	 * @generated
 	 */
@@ -137,7 +129,6 @@ public interface SCMOperator<T> extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -145,7 +136,6 @@ public interface SCMOperator<T> extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model dataType="org.enterprisedomain.classmaker.OSGiVersion"
 	 * @generated
 	 */
@@ -153,7 +143,6 @@ public interface SCMOperator<T> extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -161,7 +150,6 @@ public interface SCMOperator<T> extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -169,7 +157,6 @@ public interface SCMOperator<T> extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model exceptions="org.enterprisedomain.classmaker.Exception"
 	 * @generated
 	 */
@@ -177,7 +164,6 @@ public interface SCMOperator<T> extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model exceptions="org.enterprisedomain.classmaker.Exception"
 	 * @generated
 	 */
