+++ bundles/org.enterprisedomain.classmaker/src/org/enterprisedomain/classmaker/SCMRegistry.java
@@ -21,8 +21,9 @@ import org.eclipse.emf.ecore.EObject;
  * <!-- begin-user-doc --> A representation of the model object '<em><b>SCM
  * Registry</b></em>'. <!-- end-user-doc -->
  *
- * <!-- begin-model-doc --> Registry of Source Code Management systems. <!--
- * end-model-doc -->
+ * <!-- begin-model-doc -->
+ * Registry of Source Code Management systems.
+ * <!-- end-model-doc -->
  *
  *
  * @see org.enterprisedomain.classmaker.ClassMakerPackage#getSCMRegistry()
@@ -32,7 +33,6 @@ import org.eclipse.emf.ecore.EObject;
 public interface SCMRegistry<T> extends EObject {
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -40,7 +40,6 @@ public interface SCMRegistry<T> extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -48,7 +47,6 @@ public interface SCMRegistry<T> extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -56,7 +54,6 @@ public interface SCMRegistry<T> extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -64,7 +61,6 @@ public interface SCMRegistry<T> extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -72,7 +68,6 @@ public interface SCMRegistry<T> extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -80,7 +75,6 @@ public interface SCMRegistry<T> extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -88,7 +82,6 @@ public interface SCMRegistry<T> extends EObject {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
