+++ bundles/org.enterprisedomain.classmaker/src/org/enterprisedomain/classmaker/Stage.java
@@ -195,9 +195,9 @@ public enum Stage implements Enumerator {
 	public static final int LOADED_VALUE = 5;
 
 	/**
-	 * An array of all the '<em><b>Stage</b></em>' enumerators. <!-- begin-user-doc
+	 * An array of all the '<em><b>Stage</b></em>' enumerators.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	private static final Stage[] VALUES_ARRAY = new Stage[] { DEFINED, MODELED, GENERATED, EXPORTED, INSTALLED,
@@ -214,7 +214,6 @@ public enum Stage implements Enumerator {
 	/**
 	 * Returns the '<em><b>Stage</b></em>' literal with the specified literal value.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @param literal the literal.
 	 * @return the matching enumerator or <code>null</code>.
 	 * @generated
@@ -250,7 +249,6 @@ public enum Stage implements Enumerator {
 	/**
 	 * Returns the '<em><b>Stage</b></em>' literal with the specified integer value.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @param value the integer value.
 	 * @return the matching enumerator or <code>null</code>.
 	 * @generated
@@ -275,29 +273,26 @@ public enum Stage implements Enumerator {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	private final int value;
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	private final String name;
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	private final String literal;
 
 	/**
-	 * Only this class can construct instances. <!-- begin-user-doc --> <!--
+	 * Only this class can construct instances.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	private Stage(int value, String name, String literal) {
@@ -308,7 +303,6 @@ public enum Stage implements Enumerator {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	@Override
@@ -318,7 +312,6 @@ public enum Stage implements Enumerator {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	@Override
@@ -328,7 +321,6 @@ public enum Stage implements Enumerator {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	@Override
@@ -337,9 +329,8 @@ public enum Stage implements Enumerator {
 	}
 
 	/**
-	 * Returns the literal value of the enumerator, which is its string
-	 * representation. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns the literal value of the enumerator, which is its string representation.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	@Override
