+++ bundles/org.enterprisedomain.classmaker/src/org/enterprisedomain/classmaker/State.java
@@ -29,56 +29,37 @@ import org.eclipse.emf.ecore.resource.Resource;
  * <!-- begin-user-doc --> A representation of the model object '
  * <em><b>State</b></em>'. <!-- end-user-doc -->
  *
- * <!-- begin-model-doc --> A complete unit, representing the state of
- * <b>Contribution</b>. Identified by {@link State#getTimestamp()
- * <em>Timestamp</em>}. <!-- end-model-doc -->
+ * <!-- begin-model-doc -->
+ * A complete unit, representing the state of <b>Contribution</b>. Identified by {@link  State#getTimestamp() <em>Timestamp</em>}.
+ * <!-- end-model-doc -->
  *
  * <p>
  * The following features are supported:
  * </p>
  * <ul>
- * <li>{@link org.enterprisedomain.classmaker.State#getPackageClassName
- * <em>Package Class Name</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.State#getEditPluginClassName
- * <em>Edit Plugin Class Name</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.State#getEditorPluginClassName
- * <em>Editor Plugin Class Name</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.State#getRequiredPlugins
- * <em>Required Plugins</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.State#getRevision
- * <em>Revision</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.State#getTimestamp
- * <em>Timestamp</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.State#getDeployableUnitName
- * <em>Deployable Unit Name</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.State#getEditDeployableUnitName
- * <em>Edit Deployable Unit Name</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.State#getEditorDeployableUnitName
- * <em>Editor Deployable Unit Name</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.State#getJobFamily <em>Job
- * Family</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.State#getResource
- * <em>Resource</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.State#getCommitIds <em>Commit
- * Ids</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.State#getCommitId <em>Commit
- * Id</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.State#getStateCustomizers
- * <em>State Customizers</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.State#getProjectName <em>Project
- * Name</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.State#isMaking
- * <em>Making</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.State#isEdit <em>Edit</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.State#isEditor
- * <em>Editor</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.State#getStrategy
- * <em>Strategy</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.State#getPackageClassName <em>Package Class Name</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.State#getEditPluginClassName <em>Edit Plugin Class Name</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.State#getEditorPluginClassName <em>Editor Plugin Class Name</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.State#getRequiredPlugins <em>Required Plugins</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.State#getRevision <em>Revision</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.State#getTimestamp <em>Timestamp</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.State#getDeployableUnitName <em>Deployable Unit Name</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.State#getEditDeployableUnitName <em>Edit Deployable Unit Name</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.State#getEditorDeployableUnitName <em>Editor Deployable Unit Name</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.State#getJobFamily <em>Job Family</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.State#getResource <em>Resource</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.State#getCommitIds <em>Commit Ids</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.State#getCommitId <em>Commit Id</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.State#getStateCustomizers <em>State Customizers</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.State#getProjectName <em>Project Name</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.State#isMaking <em>Making</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.State#isEdit <em>Edit</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.State#isEditor <em>Editor</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.State#getStrategy <em>Strategy</em>}</li>
  * </ul>
  *
  * @see org.enterprisedomain.classmaker.ClassMakerPackage#getState()
- * @model superTypes="org.enterprisedomain.classmaker.Item
- *        org.enterprisedomain.classmaker.ISchedulingRule"
+ * @model superTypes="org.enterprisedomain.classmaker.Item org.enterprisedomain.classmaker.ISchedulingRule"
  * @generated
  */
 public interface State extends Item, ISchedulingRule {
@@ -101,10 +82,8 @@ public interface State extends Item, ISchedulingRule {
 	String getPackageClassName();
 
 	/**
-	 * Sets the value of the
-	 * '{@link org.enterprisedomain.classmaker.State#getPackageClassName <em>Package
-	 * Class Name</em>}' attribute. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.State#getPackageClassName <em>Package Class Name</em>}' attribute.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @param value the new value of the '<em>Package Class Name</em>' attribute.
 	 * @see #getPackageClassName()
 	 * @generated
@@ -155,13 +134,10 @@ public interface State extends Item, ISchedulingRule {
 	String getEditorPluginClassName();
 
 	/**
-	 * Sets the value of the
-	 * '{@link org.enterprisedomain.classmaker.State#getEditorPluginClassName
-	 * <em>Editor Plugin Class Name</em>}' attribute. <!-- begin-user-doc --> <!--
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.State#getEditorPluginClassName <em>Editor Plugin Class Name</em>}' attribute.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
-	 * @param value the new value of the '<em>Editor Plugin Class Name</em>'
-	 *              attribute.
+	 * @param value the new value of the '<em>Editor Plugin Class Name</em>' attribute.
 	 * @see #getEditorPluginClassName()
 	 * @generated
 	 */
@@ -203,10 +179,8 @@ public interface State extends Item, ISchedulingRule {
 	long getTimestamp();
 
 	/**
-	 * Sets the value of the
-	 * '{@link org.enterprisedomain.classmaker.State#getTimestamp
-	 * <em>Timestamp</em>}' attribute. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.State#getTimestamp <em>Timestamp</em>}' attribute.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @param value the new value of the '<em>Timestamp</em>' attribute.
 	 * @see #getTimestamp()
 	 * @generated
@@ -231,11 +205,9 @@ public interface State extends Item, ISchedulingRule {
 	String getDeployableUnitName();
 
 	/**
-	 * Sets the value of the
-	 * '{@link org.enterprisedomain.classmaker.State#getDeployableUnitName
-	 * <em>Deployable Unit Name</em>}' attribute. <!-- begin-user-doc --> <!--
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.State#getDeployableUnitName <em>Deployable Unit Name</em>}' attribute.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @param value the new value of the '<em>Deployable Unit Name</em>' attribute.
 	 * @see #getDeployableUnitName()
 	 * @generated
@@ -257,13 +229,10 @@ public interface State extends Item, ISchedulingRule {
 	String getEditDeployableUnitName();
 
 	/**
-	 * Sets the value of the
-	 * '{@link org.enterprisedomain.classmaker.State#getEditDeployableUnitName
-	 * <em>Edit Deployable Unit Name</em>}' attribute. <!-- begin-user-doc --> <!--
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.State#getEditDeployableUnitName <em>Edit Deployable Unit Name</em>}' attribute.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
-	 * @param value the new value of the '<em>Edit Deployable Unit Name</em>'
-	 *              attribute.
+	 * @param value the new value of the '<em>Edit Deployable Unit Name</em>' attribute.
 	 * @see #getEditDeployableUnitName()
 	 * @generated
 	 */
@@ -284,13 +253,10 @@ public interface State extends Item, ISchedulingRule {
 	String getEditorDeployableUnitName();
 
 	/**
-	 * Sets the value of the
-	 * '{@link org.enterprisedomain.classmaker.State#getEditorDeployableUnitName
-	 * <em>Editor Deployable Unit Name</em>}' attribute. <!-- begin-user-doc -->
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.State#getEditorDeployableUnitName <em>Editor Deployable Unit Name</em>}' attribute.
+	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
-	 * 
-	 * @param value the new value of the '<em>Editor Deployable Unit Name</em>'
-	 *              attribute.
+	 * @param value the new value of the '<em>Editor Deployable Unit Name</em>' attribute.
 	 * @see #getEditorDeployableUnitName()
 	 * @generated
 	 */
@@ -314,10 +280,8 @@ public interface State extends Item, ISchedulingRule {
 	Revision getRevision();
 
 	/**
-	 * Sets the value of the
-	 * '{@link org.enterprisedomain.classmaker.State#getRevision <em>Revision</em>}'
-	 * reference. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.State#getRevision <em>Revision</em>}' reference.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @param value the new value of the '<em>Revision</em>' reference.
 	 * @see #getRevision()
 	 * @generated
@@ -342,10 +306,8 @@ public interface State extends Item, ISchedulingRule {
 	String getJobFamily();
 
 	/**
-	 * Sets the value of the
-	 * '{@link org.enterprisedomain.classmaker.State#getJobFamily <em>Job
-	 * Family</em>}' attribute. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.State#getJobFamily <em>Job Family</em>}' attribute.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @param value the new value of the '<em>Job Family</em>' attribute.
 	 * @see #getJobFamily()
 	 * @generated
@@ -370,10 +332,8 @@ public interface State extends Item, ISchedulingRule {
 	Resource getResource();
 
 	/**
-	 * Sets the value of the
-	 * '{@link org.enterprisedomain.classmaker.State#getResource <em>Resource</em>}'
-	 * reference. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.State#getResource <em>Resource</em>}' reference.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @param value the new value of the '<em>Resource</em>' reference.
 	 * @see #getResource()
 	 * @generated
@@ -416,10 +376,8 @@ public interface State extends Item, ISchedulingRule {
 	String getCommitId();
 
 	/**
-	 * Sets the value of the
-	 * '{@link org.enterprisedomain.classmaker.State#getCommitId <em>Commit
-	 * Id</em>}' attribute. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.State#getCommitId <em>Commit Id</em>}' attribute.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @param value the new value of the '<em>Commit Id</em>' attribute.
 	 * @see #getCommitId()
 	 * @generated
@@ -465,10 +423,8 @@ public interface State extends Item, ISchedulingRule {
 	String getProjectName();
 
 	/**
-	 * Sets the value of the
-	 * '{@link org.enterprisedomain.classmaker.State#getProjectName <em>Project
-	 * Name</em>}' attribute. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.State#getProjectName <em>Project Name</em>}' attribute.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @param value the new value of the '<em>Project Name</em>' attribute.
 	 * @see #getProjectName()
 	 * @generated
@@ -494,9 +450,8 @@ public interface State extends Item, ISchedulingRule {
 	boolean isMaking();
 
 	/**
-	 * Sets the value of the '{@link org.enterprisedomain.classmaker.State#isMaking
-	 * <em>Making</em>}' attribute. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.State#isMaking <em>Making</em>}' attribute.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @param value the new value of the '<em>Making</em>' attribute.
 	 * @see #isMaking()
 	 * @generated
@@ -516,9 +471,8 @@ public interface State extends Item, ISchedulingRule {
 	boolean isEdit();
 
 	/**
-	 * Sets the value of the '{@link org.enterprisedomain.classmaker.State#isEdit
-	 * <em>Edit</em>}' attribute. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.State#isEdit <em>Edit</em>}' attribute.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @param value the new value of the '<em>Edit</em>' attribute.
 	 * @see #isEdit()
 	 * @generated
@@ -538,9 +492,8 @@ public interface State extends Item, ISchedulingRule {
 	boolean isEditor();
 
 	/**
-	 * Sets the value of the '{@link org.enterprisedomain.classmaker.State#isEditor
-	 * <em>Editor</em>}' attribute. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.State#isEditor <em>Editor</em>}' attribute.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @param value the new value of the '<em>Editor</em>' attribute.
 	 * @see #isEditor()
 	 * @generated
@@ -548,11 +501,9 @@ public interface State extends Item, ISchedulingRule {
 	void setEditor(boolean value);
 
 	/**
-	 * Returns the value of the '<em><b>Strategy</b></em>' reference. It is
-	 * bidirectional and its opposite is
-	 * '{@link org.enterprisedomain.classmaker.Strategy#getState <em>State</em>}'.
+	 * Returns the value of the '<em><b>Strategy</b></em>' reference.
+	 * It is bidirectional and its opposite is '{@link org.enterprisedomain.classmaker.Strategy#getState <em>State</em>}'.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @return the value of the '<em>Strategy</em>' reference.
 	 * @see #setStrategy(Strategy)
 	 * @see org.enterprisedomain.classmaker.ClassMakerPackage#getState_Strategy()
@@ -563,10 +514,8 @@ public interface State extends Item, ISchedulingRule {
 	Strategy getStrategy();
 
 	/**
-	 * Sets the value of the
-	 * '{@link org.enterprisedomain.classmaker.State#getStrategy <em>Strategy</em>}'
-	 * reference. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.State#getStrategy <em>Strategy</em>}' reference.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @param value the new value of the '<em>Strategy</em>' reference.
 	 * @see #getStrategy()
 	 * @generated
@@ -575,25 +524,20 @@ public interface State extends Item, ISchedulingRule {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model exceptions="org.enterprisedomain.classmaker.CoreException"
-	 *        monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
+	 * @model exceptions="org.enterprisedomain.classmaker.CoreException" monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
 	 * @generated
 	 */
 	void setProjectVersion(IProgressMonitor monitor) throws CoreException;
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model exceptions="org.enterprisedomain.classmaker.CoreException"
-	 *        monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
+	 * @model exceptions="org.enterprisedomain.classmaker.CoreException" monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
 	 * @generated
 	 */
 	void delete(IProgressMonitor monitor) throws CoreException;
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -601,7 +545,6 @@ public interface State extends Item, ISchedulingRule {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -609,7 +552,6 @@ public interface State extends Item, ISchedulingRule {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -617,7 +559,6 @@ public interface State extends Item, ISchedulingRule {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model exceptions="org.enterprisedomain.classmaker.Exception"
 	 * @generated
 	 */
@@ -635,7 +576,6 @@ public interface State extends Item, ISchedulingRule {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
