+++ bundles/org.enterprisedomain.classmaker/src/org/enterprisedomain/classmaker/Workspace.java
@@ -28,25 +28,19 @@ import org.eclipse.emf.ecore.resource.ResourceSet;
  * <!-- begin-user-doc --> A representation of the model object '
  * <em><b>Workspace</b></em>'. <!-- end-user-doc -->
  *
- * <!-- begin-model-doc --> Maps
- * {@link org.eclipse.emf.ecore.resource.ResourceSet ResourceSet} to
- * {@link org.eclipse.core.resources.IWorkspaceRoot IWorkspaceRoot}. <!--
- * end-model-doc -->
+ * <!-- begin-model-doc -->
+ * Maps {@link org.eclipse.emf.ecore.resource.ResourceSet ResourceSet} to {@link org.eclipse.core.resources.IWorkspaceRoot IWorkspaceRoot}.
+ * <!-- end-model-doc -->
  *
  * <p>
  * The following features are supported:
  * </p>
  * <ul>
- * <li>{@link org.enterprisedomain.classmaker.Workspace#getProjects
- * <em>Projects</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.Workspace#getResourceSet
- * <em>Resource Set</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.Workspace#getCustomizers
- * <em>Customizers</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.Workspace#getService
- * <em>Service</em>}</li>
- * <li>{@link org.enterprisedomain.classmaker.Workspace#getSCMRegistry <em>SCM
- * Registry</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.Workspace#getProjects <em>Projects</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.Workspace#getResourceSet <em>Resource Set</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.Workspace#getCustomizers <em>Customizers</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.Workspace#getService <em>Service</em>}</li>
+ *   <li>{@link org.enterprisedomain.classmaker.Workspace#getSCMRegistry <em>SCM Registry</em>}</li>
  * </ul>
  *
  * @see org.enterprisedomain.classmaker.ClassMakerPackage#getWorkspace()
@@ -108,16 +102,14 @@ public interface Workspace extends EObject, ISchedulingRule {
 	EMap<StageQualifier, Customizer> getCustomizers();
 
 	/**
-	 * Returns the value of the '<em><b>Service</b></em>' container reference. It is
-	 * bidirectional and its opposite is
-	 * '{@link org.enterprisedomain.classmaker.ClassMakerService#getWorkspace
-	 * <em>Workspace</em>}'. <!-- begin-user-doc -->
+	 * Returns the value of the '<em><b>Service</b></em>' container reference.
+	 * It is bidirectional and its opposite is '{@link org.enterprisedomain.classmaker.ClassMakerService#getWorkspace <em>Workspace</em>}'.
+	 * <!-- begin-user-doc -->
 	 * <p>
 	 * If the meaning of the '<em>Service</em>' reference isn't clear, there really
 	 * should be more of a description here...
 	 * </p>
 	 * <!-- end-user-doc -->
-	 * 
 	 * @return the value of the '<em>Service</em>' container reference.
 	 * @see #setService(ClassMakerService)
 	 * @see org.enterprisedomain.classmaker.ClassMakerPackage#getWorkspace_Service()
@@ -128,11 +120,9 @@ public interface Workspace extends EObject, ISchedulingRule {
 	ClassMakerService getService();
 
 	/**
-	 * Sets the value of the
-	 * '{@link org.enterprisedomain.classmaker.Workspace#getService
-	 * <em>Service</em>}' container reference. <!-- begin-user-doc --> <!--
+	 * Sets the value of the '{@link org.enterprisedomain.classmaker.Workspace#getService <em>Service</em>}' container reference.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @param value the new value of the '<em>Service</em>' container reference.
 	 * @see #getService()
 	 * @generated
@@ -282,7 +272,6 @@ public interface Workspace extends EObject, ISchedulingRule {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -290,7 +279,6 @@ public interface Workspace extends EObject, ISchedulingRule {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -298,7 +286,6 @@ public interface Workspace extends EObject, ISchedulingRule {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @model
 	 * @generated
 	 */
@@ -306,9 +293,7 @@ public interface Workspace extends EObject, ISchedulingRule {
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
-	 * @model exceptions="org.enterprisedomain.classmaker.CoreException"
-	 *        monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
+	 * @model exceptions="org.enterprisedomain.classmaker.CoreException" monitorDataType="org.enterprisedomain.classmaker.IProgressMonitor"
 	 * @generated
 	 */
 	void delete(Object object, IProgressMonitor monitor) throws CoreException;
