+++ bundles/org.enterprisedomain.editor/src/org/enterprisedomain/classmaker/presentation/ClassMakerActionBarContributor.java
@@ -63,9 +63,9 @@ import org.eclipse.ui.PartInitException;
 public class ClassMakerActionBarContributor extends EditingDomainActionBarContributor
 		implements ISelectionChangedListener {
 	/**
-	 * Action to create objects from the ClassMaker model. <!-- begin-user-doc -->
+	 * Action to create objects from the ClassMaker model.
+	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	public static class NewAction extends WorkbenchWindowActionDelegate {
@@ -84,25 +84,25 @@ public class ClassMakerActionBarContributor extends EditingDomainActionBarContri
 	}
 
 	/**
-	 * This keeps track of the active editor. <!-- begin-user-doc --> <!--
+	 * This keeps track of the active editor.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	protected IEditorPart activeEditorPart;
 
 	/**
-	 * This keeps track of the current selection provider. <!-- begin-user-doc -->
+	 * This keeps track of the current selection provider.
+	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	protected ISelectionProvider selectionProvider;
 
 	/**
-	 * This action opens the Properties view. <!-- begin-user-doc --> <!--
+	 * This action opens the Properties view.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	protected IAction showPropertiesViewAction = new Action(
@@ -143,63 +143,53 @@ public class ClassMakerActionBarContributor extends EditingDomainActionBarContri
 	};
 
 	/**
-	 * This will contain one
-	 * {@link org.eclipse.emf.edit.ui.action.CreateChildAction} corresponding to
-	 * each descriptor generated for the current selection by the item provider.
+	 * This will contain one {@link org.eclipse.emf.edit.ui.action.CreateChildAction} corresponding to each descriptor
+	 * generated for the current selection by the item provider.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	protected Collection<IAction> createChildActions;
 
 	/**
-	 * This will contain a map of
-	 * {@link org.eclipse.emf.edit.ui.action.CreateChildAction}s, keyed by sub-menu
-	 * text. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * This will contain a map of {@link org.eclipse.emf.edit.ui.action.CreateChildAction}s, keyed by sub-menu text.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	protected Map<String, Collection<IAction>> createChildSubmenuActions;
 
 	/**
-	 * This is the menu manager into which menu contribution items should be added
-	 * for CreateChild actions. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * This is the menu manager into which menu contribution items should be added for CreateChild actions.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	protected IMenuManager createChildMenuManager;
 
 	/**
-	 * This will contain one
-	 * {@link org.eclipse.emf.edit.ui.action.CreateSiblingAction} corresponding to
-	 * each descriptor generated for the current selection by the item provider.
+	 * This will contain one {@link org.eclipse.emf.edit.ui.action.CreateSiblingAction} corresponding to each descriptor
+	 * generated for the current selection by the item provider.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	protected Collection<IAction> createSiblingActions;
 
 	/**
-	 * This will contain a map of
-	 * {@link org.eclipse.emf.edit.ui.action.CreateSiblingAction}s, keyed by submenu
-	 * text. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * This will contain a map of {@link org.eclipse.emf.edit.ui.action.CreateSiblingAction}s, keyed by submenu text.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	protected Map<String, Collection<IAction>> createSiblingSubmenuActions;
 
 	/**
-	 * This is the menu manager into which menu contribution items should be added
-	 * for CreateSibling actions. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * This is the menu manager into which menu contribution items should be added for CreateSibling actions.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	protected IMenuManager createSiblingMenuManager;
 
 	/**
-	 * This creates an instance of the contributor. <!-- begin-user-doc --> <!--
+	 * This creates an instance of the contributor.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	public ClassMakerActionBarContributor() {
@@ -223,10 +213,10 @@ public class ClassMakerActionBarContributor extends EditingDomainActionBarContri
 	}
 
 	/**
-	 * This adds to the menu bar a menu and some separators for editor additions, as
-	 * well as the sub-menus for object creation items. <!-- begin-user-doc --> <!--
+	 * This adds to the menu bar a menu and some separators for editor additions,
+	 * as well as the sub-menus for object creation items.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	@Override
@@ -267,9 +257,8 @@ public class ClassMakerActionBarContributor extends EditingDomainActionBarContri
 	}
 
 	/**
-	 * When the active editor changes, this remembers the change and registers with
-	 * it as a selection provider. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * When the active editor changes, this remembers the change and registers with it as a selection provider.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	@Override
@@ -298,10 +287,9 @@ public class ClassMakerActionBarContributor extends EditingDomainActionBarContri
 
 	/**
 	 * This implements {@link org.eclipse.jface.viewers.ISelectionChangedListener},
-	 * handling {@link org.eclipse.jface.viewers.SelectionChangedEvent}s by querying
-	 * for the children and siblings that can be added to the selected object and
-	 * updating the menus accordingly. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * handling {@link org.eclipse.jface.viewers.SelectionChangedEvent}s by querying for the children and siblings
+	 * that can be added to the selected object and updating the menus accordingly.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	@Override
@@ -352,10 +340,9 @@ public class ClassMakerActionBarContributor extends EditingDomainActionBarContri
 	}
 
 	/**
-	 * This generates a {@link org.eclipse.emf.edit.ui.action.CreateChildAction} for
-	 * each object in <code>descriptors</code>, and returns the collection of these
-	 * actions. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * This generates a {@link org.eclipse.emf.edit.ui.action.CreateChildAction} for each object in <code>descriptors</code>,
+	 * and returns the collection of these actions.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	protected Collection<IAction> generateCreateChildActions(Collection<?> descriptors, ISelection selection) {
@@ -369,10 +356,9 @@ public class ClassMakerActionBarContributor extends EditingDomainActionBarContri
 	}
 
 	/**
-	 * This generates a {@link org.eclipse.emf.edit.ui.action.CreateSiblingAction}
-	 * for each object in <code>descriptors</code>, and returns the collection of
-	 * these actions. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * This generates a {@link org.eclipse.emf.edit.ui.action.CreateSiblingAction} for each object in <code>descriptors</code>,
+	 * and returns the collection of these actions.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	protected Collection<IAction> generateCreateSiblingActions(Collection<?> descriptors, ISelection selection) {
@@ -386,14 +372,12 @@ public class ClassMakerActionBarContributor extends EditingDomainActionBarContri
 	}
 
 	/**
-	 * This populates the specified <code>manager</code> with
-	 * {@link org.eclipse.jface.action.ActionContributionItem}s based on the
-	 * {@link org.eclipse.jface.action.IAction}s contained in the
-	 * <code>actions</code> collection, by inserting them before the specified
-	 * contribution item <code>contributionID</code>. If <code>contributionID</code>
-	 * is <code>null</code>, they are simply added. <!-- begin-user-doc --> <!--
+	 * This populates the specified <code>manager</code> with {@link org.eclipse.jface.action.ActionContributionItem}s
+	 * based on the {@link org.eclipse.jface.action.IAction}s contained in the <code>actions</code> collection,
+	 * by inserting them before the specified contribution item <code>contributionID</code>.
+	 * If <code>contributionID</code> is <code>null</code>, they are simply added.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	protected void populateManager(IContributionManager manager, Collection<? extends IAction> actions,
@@ -442,10 +426,9 @@ public class ClassMakerActionBarContributor extends EditingDomainActionBarContri
 	}
 
 	/**
-	 * This extracts those actions in the <code>submenuActions</code> collection
-	 * whose text is qualified and returns a map of these actions, keyed by submenu
-	 * text. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * This extracts those actions in the <code>submenuActions</code> collection whose text is qualified and returns
+	 * a map of these actions, keyed by submenu text.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	protected Map<String, Collection<IAction>> extractSubmenuActions(Collection<IAction> createActions) {
@@ -523,9 +506,9 @@ public class ClassMakerActionBarContributor extends EditingDomainActionBarContri
 	}
 
 	/**
-	 * This populates the pop-up menu before it appears. <!-- begin-user-doc -->
+	 * This populates the pop-up menu before it appears.
+	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	@Override
@@ -562,9 +545,8 @@ public class ClassMakerActionBarContributor extends EditingDomainActionBarContri
 	}
 
 	/**
-	 * This ensures that a delete action will clean up all references to deleted
-	 * objects. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * This ensures that a delete action will clean up all references to deleted objects.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	@Override
