+++ bundles/org.enterprisedomain.editor/src/org/enterprisedomain/classmaker/presentation/ClassMakerEditor.java
@@ -111,26 +111,25 @@ import org.eclipse.ui.views.properties.PropertySheetPage;
 import org.enterprisedomain.classmaker.provider.ClassMakerItemProviderAdapterFactory;
 
 /**
- * This is an example of a ClassMaker model editor. <!-- begin-user-doc --> <!--
+ * This is an example of a ClassMaker model editor.
+ * <!-- begin-user-doc --> <!--
  * end-user-doc -->
- * 
  * @generated
  */
 public class ClassMakerEditor extends MultiPageEditorPart
 		implements IEditingDomainProvider, ISelectionProvider, IMenuListener, IViewerProvider {
 	/**
-	 * The filters for file extensions supported by the editor. <!-- begin-user-doc
+	 * The filters for file extensions supported by the editor.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	public static final List<String> FILE_EXTENSION_FILTERS = prefixExtensions(ClassMakerModelWizard.FILE_EXTENSIONS,
 			"*.");
 
 	/**
-	 * Returns a new unmodifiable list containing prefixed versions of the
-	 * extensions in the given list. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Returns a new unmodifiable list containing prefixed versions of the extensions in the given list.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	private static List<String> prefixExtensions(List<String> extensions, String prefix) {
@@ -142,9 +141,8 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	}
 
 	/**
-	 * This keeps track of the editing domain that is used to track all changes to
-	 * the model. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * This keeps track of the editing domain that is used to track all changes to the model.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	protected AdapterFactoryEditingDomain editingDomain;
@@ -166,16 +164,16 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	protected IContentOutlinePage contentOutlinePage;
 
 	/**
-	 * This is a kludge... <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * This is a kludge...
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	protected IStatusLineManager contentOutlineStatusLineManager;
 
 	/**
-	 * This is the content outline page's viewer. <!-- begin-user-doc --> <!--
+	 * This is the content outline page's viewer.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	protected TreeViewer contentOutlineViewer;
@@ -198,27 +196,24 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	protected TreeViewer selectionViewer;
 
 	/**
-	 * This keeps track of the active content viewer, which may be either one of the
-	 * viewers in the pages or the content outline viewer. <!-- begin-user-doc -->
+	 * This keeps track of the active content viewer, which may be either one of the viewers in the pages or the content outline viewer.
+	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	protected Viewer currentViewer;
 
 	/**
-	 * This listens to which ever viewer is active. <!-- begin-user-doc --> <!--
+	 * This listens to which ever viewer is active.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	protected ISelectionChangedListener selectionChangedListener;
 
 	/**
-	 * This keeps track of all the
-	 * {@link org.eclipse.jface.viewers.ISelectionChangedListener}s that are
-	 * listening to this editor. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * This keeps track of all the {@link org.eclipse.jface.viewers.ISelectionChangedListener}s that are listening to this editor.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	protected Collection<ISelectionChangedListener> selectionChangedListeners = new ArrayList<ISelectionChangedListener>();
@@ -232,9 +227,9 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	protected ISelection editorSelection = StructuredSelection.EMPTY;
 
 	/**
-	 * This listens for when the outline becomes active <!-- begin-user-doc --> <!--
+	 * This listens for when the outline becomes active
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	protected IPartListener partListener = new IPartListener() {
@@ -278,32 +273,32 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	};
 
 	/**
-	 * Resources that have been removed since last activation. <!-- begin-user-doc
+	 * Resources that have been removed since last activation.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	protected Collection<Resource> removedResources = new ArrayList<Resource>();
 
 	/**
-	 * Resources that have been changed since last activation. <!-- begin-user-doc
+	 * Resources that have been changed since last activation.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	protected Collection<Resource> changedResources = new ArrayList<Resource>();
 
 	/**
-	 * Resources that have been saved. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Resources that have been saved.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	protected Collection<Resource> savedResources = new ArrayList<Resource>();
 
 	/**
-	 * Map to store the diagnostic associated with a resource. <!-- begin-user-doc
+	 * Map to store the diagnostic associated with a resource.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	protected Map<Resource, Diagnostic> resourceToDiagnosticMap = new LinkedHashMap<Resource, Diagnostic>();
@@ -317,9 +312,8 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	protected boolean updateProblemIndication = true;
 
 	/**
-	 * Adapter used to update the problem indication when resources are demanded
-	 * loaded. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Adapter used to update the problem indication when resources are demanded loaded.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	protected EContentAdapter problemIndicationAdapter = new EContentAdapter() {
@@ -408,9 +402,9 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	}
 
 	/**
-	 * Handles what to do with changed resources on activation. <!-- begin-user-doc
+	 * Handles what to do with changed resources on activation.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	protected void handleChangedResources() {
@@ -445,9 +439,8 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	}
 
 	/**
-	 * Updates the problems indication with the information described in the
-	 * specified diagnostic. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * Updates the problems indication with the information described in the specified diagnostic.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	protected void updateProblemIndication() {
@@ -493,8 +486,8 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	}
 
 	/**
-	 * This creates a model editor. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * This creates a model editor.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	public ClassMakerEditor() {
@@ -503,9 +496,9 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	}
 
 	/**
-	 * This sets up the editing domain for the model editor. <!-- begin-user-doc -->
+	 * This sets up the editing domain for the model editor.
+	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	protected void initializeEditingDomain() {
@@ -518,13 +511,11 @@ public class ClassMakerEditor extends MultiPageEditorPart
 		adapterFactory.addAdapterFactory(new EcoreItemProviderAdapterFactory());
 		adapterFactory.addAdapterFactory(new ReflectiveItemProviderAdapterFactory());
 
-		// Create the command stack that will notify this editor as commands are
-		// executed.
+		// Create the command stack that will notify this editor as commands are executed.
 		//
 		BasicCommandStack commandStack = new BasicCommandStack();
 
-		// Add a listener to set the most recent command's affected objects to be the
-		// selection of the viewer with focus.
+		// Add a listener to set the most recent command's affected objects to be the selection of the viewer with focus.
 		//
 		commandStack.addCommandStackListener(new CommandStackListener() {
 			@Override
@@ -559,9 +550,9 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	}
 
 	/**
-	 * This is here for the listener to be able to call it. <!-- begin-user-doc -->
+	 * This is here for the listener to be able to call it.
+	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	@Override
@@ -570,9 +561,9 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	}
 
 	/**
-	 * This sets the selection into whichever viewer is active. <!-- begin-user-doc
+	 * This sets the selection into whichever viewer is active.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	public void setSelectionToViewer(Collection<?> collection) {
@@ -595,12 +586,11 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	}
 
 	/**
-	 * This returns the editing domain as required by the
-	 * {@link IEditingDomainProvider} interface. This is important for implementing
-	 * the static methods of {@link AdapterFactoryEditingDomain} and for supporting
-	 * {@link org.eclipse.emf.edit.ui.action.CommandAction}. <!-- begin-user-doc -->
+	 * This returns the editing domain as required by the {@link IEditingDomainProvider} interface.
+	 * This is important for implementing the static methods of {@link AdapterFactoryEditingDomain}
+	 * and for supporting {@link org.eclipse.emf.edit.ui.action.CommandAction}.
+	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	@Override
@@ -610,13 +600,11 @@ public class ClassMakerEditor extends MultiPageEditorPart
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	public class ReverseAdapterFactoryContentProvider extends AdapterFactoryContentProvider {
 		/**
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		public ReverseAdapterFactoryContentProvider(AdapterFactory adapterFactory) {
@@ -625,7 +613,6 @@ public class ClassMakerEditor extends MultiPageEditorPart
 
 		/**
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		@Override
@@ -636,7 +623,6 @@ public class ClassMakerEditor extends MultiPageEditorPart
 
 		/**
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		@Override
@@ -647,7 +633,6 @@ public class ClassMakerEditor extends MultiPageEditorPart
 
 		/**
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		@Override
@@ -658,7 +643,6 @@ public class ClassMakerEditor extends MultiPageEditorPart
 
 		/**
 		 * <!-- begin-user-doc --> <!-- end-user-doc -->
-		 * 
 		 * @generated
 		 */
 		@Override
@@ -668,10 +652,10 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	}
 
 	/**
-	 * This makes sure that one content viewer, either for the current page or the
-	 * outline view, if it has focus, is the current one. <!-- begin-user-doc -->
+	 * This makes sure that one content viewer, either for the current page or the outline view, if it has focus,
+	 * is the current one.
+	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	public void setCurrentViewer(Viewer viewer) {
@@ -716,7 +700,6 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	/**
 	 * This returns the viewer as required by the {@link IViewerProvider} interface.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	@Override
@@ -748,10 +731,9 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	}
 
 	/**
-	 * This is the method called to load a resource into the editing domain's
-	 * resource set based on the editor's input. <!-- begin-user-doc --> <!--
+	 * This is the method called to load a resource into the editing domain's resource set based on the editor's input.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	public void createModel() {
@@ -775,10 +757,10 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	}
 
 	/**
-	 * Returns a diagnostic describing the errors and warnings listed in the
-	 * resource and the specified exception (if any). <!-- begin-user-doc --> <!--
+	 * Returns a diagnostic describing the errors and warnings listed in the resource
+	 * and the specified exception (if any).
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	public Diagnostic analyzeResourceProblems(Resource resource, Exception exception) {
@@ -866,9 +848,9 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	}
 
 	/**
-	 * If there is just one page in the multi-page editor part, this hides the
-	 * single tab at the bottom. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * If there is just one page in the multi-page editor part,
+	 * this hides the single tab at the bottom.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	protected void hideTabs() {
@@ -883,9 +865,9 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	}
 
 	/**
-	 * If there is more than one page in the multi-page editor part, this shows the
-	 * tabs at the bottom. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * If there is more than one page in the multi-page editor part,
+	 * this shows the tabs at the bottom.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	protected void showTabs() {
@@ -900,9 +882,9 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	}
 
 	/**
-	 * This is used to track the active viewer. <!-- begin-user-doc --> <!--
+	 * This is used to track the active viewer.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	@Override
@@ -933,9 +915,9 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	}
 
 	/**
-	 * This accesses a cached version of the content outliner. <!-- begin-user-doc
+	 * This accesses a cached version of the content outliner.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	public IContentOutlinePage getContentOutlinePage() {
@@ -1000,9 +982,9 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	}
 
 	/**
-	 * This accesses a cached version of the property sheet. <!-- begin-user-doc -->
+	 * This accesses a cached version of the property sheet.
+	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	public IPropertySheetPage getPropertySheetPage() {
@@ -1027,9 +1009,8 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	}
 
 	/**
-	 * This deals with how we want selection in the outliner to affect the other
-	 * views. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * This deals with how we want selection in the outliner to affect the other views.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	public void handleContentOutlineSelection(ISelection selection) {
@@ -1054,9 +1035,8 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	}
 
 	/**
-	 * This is for implementing {@link IEditorPart} and simply tests the command
-	 * stack. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * This is for implementing {@link IEditorPart} and simply tests the command stack.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	@Override
@@ -1067,7 +1047,6 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	/**
 	 * This is for implementing {@link IEditorPart} and simply saves the model file.
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	@Override
@@ -1078,8 +1057,7 @@ public class ClassMakerEditor extends MultiPageEditorPart
 		saveOptions.put(Resource.OPTION_SAVE_ONLY_IF_CHANGED, Resource.OPTION_SAVE_ONLY_IF_CHANGED_MEMORY_BUFFER);
 		saveOptions.put(Resource.OPTION_LINE_DELIMITER, Resource.OPTION_LINE_DELIMITER_UNSPECIFIED);
 
-		// Do the work within an operation because this is a long running activity that
-		// modifies the workbench.
+		// Do the work within an operation because this is a long running activity that modifies the workbench.
 		//
 		IRunnableWithProgress operation = new IRunnableWithProgress() {
 			// This is the method that gets invoked when the operation runs.
@@ -1129,11 +1107,10 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	}
 
 	/**
-	 * This returns whether something has been persisted to the URI of the specified
-	 * resource. The implementation uses the URI converter from the editor's
-	 * resource set to try to open an input stream. <!-- begin-user-doc --> <!--
+	 * This returns whether something has been persisted to the URI of the specified resource.
+	 * The implementation uses the URI converter from the editor's resource set to try to open an input stream.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	protected boolean isPersisted(Resource resource) {
@@ -1162,9 +1139,9 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	}
 
 	/**
-	 * This also changes the editor's input. <!-- begin-user-doc --> <!--
+	 * This also changes the editor's input.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	@Override
@@ -1179,7 +1156,6 @@ public class ClassMakerEditor extends MultiPageEditorPart
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	protected void doSaveAs(URI uri, IEditorInput editorInput) {
@@ -1193,8 +1169,8 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	}
 
 	/**
-	 * This is called during startup. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * This is called during startup.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	@Override
@@ -1208,7 +1184,6 @@ public class ClassMakerEditor extends MultiPageEditorPart
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	@Override
@@ -1239,10 +1214,9 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	}
 
 	/**
-	 * This implements {@link org.eclipse.jface.viewers.ISelectionProvider} to
-	 * return this editor's overall selection. <!-- begin-user-doc --> <!--
+	 * This implements {@link org.eclipse.jface.viewers.ISelectionProvider} to return this editor's overall selection.
+	 * <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	@Override
@@ -1251,10 +1225,9 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	}
 
 	/**
-	 * This implements {@link org.eclipse.jface.viewers.ISelectionProvider} to set
-	 * this editor's overall selection. Calling this result will notify the
-	 * listeners. <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
+	 * This implements {@link org.eclipse.jface.viewers.ISelectionProvider} to set this editor's overall selection.
+	 * Calling this result will notify the listeners.
+	 * <!-- begin-user-doc --> <!-- end-user-doc -->
 	 * @generated
 	 */
 	@Override
@@ -1269,7 +1242,6 @@ public class ClassMakerEditor extends MultiPageEditorPart
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	public void setStatusLineManager(ISelection selection) {
@@ -1323,10 +1295,9 @@ public class ClassMakerEditor extends MultiPageEditorPart
 	}
 
 	/**
-	 * This implements {@link org.eclipse.jface.action.IMenuListener} to help fill
-	 * the context menus with contributions from the Edit menu. <!-- begin-user-doc
+	 * This implements {@link org.eclipse.jface.action.IMenuListener} to help fill the context menus with contributions from the Edit menu.
+	 * <!-- begin-user-doc
 	 * --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	@Override
@@ -1336,7 +1307,6 @@ public class ClassMakerEditor extends MultiPageEditorPart
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	public EditingDomainActionBarContributor getActionBarContributor() {
@@ -1345,7 +1315,6 @@ public class ClassMakerEditor extends MultiPageEditorPart
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	public IActionBars getActionBars() {
@@ -1354,7 +1323,6 @@ public class ClassMakerEditor extends MultiPageEditorPart
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	public AdapterFactory getAdapterFactory() {
@@ -1363,7 +1331,6 @@ public class ClassMakerEditor extends MultiPageEditorPart
 
 	/**
 	 * <!-- begin-user-doc --> <!-- end-user-doc -->
-	 * 
 	 * @generated
 	 */
 	@Override
