+++ tests/org.enterprisedomain.classmaker.tests/META-INF/MANIFEST.MF
@@ -13,7 +13,9 @@ Require-Bundle: org.eclipse.core.runtime,
  org.eclipse.emf.ecore;visibility:=reexport,
  org.eclipse.emf.ecore.xmi;visibility:=reexport,
  org.eclipse.pde.junit.runtime;bundle-version="[3.5.50,4.0.0)",
- org.junit;bundle-version="[4.12.0,5.0.0)";visibility:=reexport
+ org.junit,
+ org.junit.jupiter.api,
+ org.eclipse.jdt.junit5.runtime;bundle-version="1.0.1100"
 Bundle-RequiredExecutionEnvironment: JavaSE-11
 Export-Package: org.enterprisedomain.classmaker.tests
 Service-Component: OSGI-INF/component.xml
